+++ src/appLogic.js
@@ -1,0 +1,2739 @@
+import React, { useState, useRef, useEffect } from 'react';
+import GoogleSheetsService from './services/googleSheetsService';
+import GoogleDriveService from './services/googleDriveService';
+import GoogleAuthService from './services/googleAuthService';
+
+export const SECTION_LIST = [
+  'main', 'drive', 'portal', 'pptMaker', 'myPage'
+];
+
+function useAppLogic() {
+    // 상태 관리
+    const [isLoggedIn, setIsLoggedIn] = useState(false);
+    const [activeSection, setActiveSection] = useState(() => {
+      // localStorage에서 저장된 섹션 복원
+      return localStorage.getItem('activeSection') || 'main';
+    });
+    // App.js 상단 상태 관리 섹션에 추가
+    const [originalPeriod, setOriginalPeriod] = useState(null); // 기존 설정 기간을 저장
+    const [showModal, setShowModal] = useState(false);
+    const [experiences, setExperiences] = useState([]);
+    const [form, setForm] = useState({ title: '', startDate: '', endDate: '', description: '' });
+    const [selected, setSelected] = useState([]);
+    const [spreadsheetId, setSpreadsheetId] = useState(() => {
+      // localStorage에서 스프레드시트 ID 복원
+      return localStorage.getItem('spreadsheetId') || null;
+    });
+    const [isSheetsInitialized, setIsSheetsInitialized] = useState(false);
+    const [isDriveInitialized, setIsDriveInitialized] = useState(false);
+    const [driveFiles, setDriveFiles] = useState([]);
+    const [isLoading, setIsLoading] = useState(false);
+    const [isSheetLoading, setIsSheetLoading] = useState(false);
+    const [isExperienceLoading, setIsExperienceLoading] = useState(false);
+    const [isDriveLoading, setIsDriveLoading] = useState(false);
+    const [isUploadLoading, setIsUploadLoading] = useState(false);
+    const [isRefreshLoading, setIsRefreshLoading] = useState(false);
+    const [isDeleteLoading, setIsDeleteLoading] = useState(false);
+    const [editingIndex, setEditingIndex] = useState(null);
+    const [deletingFileIds, setDeletingFileIds] = useState(new Set()); // 삭제 중인 파일 ID들
+    const [isViewModeLoading, setIsViewModeLoading] = useState(false);
+    const [isPptCreating, setIsPptCreating] = useState(false); // PPT 생성 로딩 상태
+    const [currentPath, setCurrentPath] = useState([]); // 현재 경로 추적
+    const [authStatus, setAuthStatus] = useState('disconnected');
+    const [selectedImages, setSelectedImages] = useState([]); // 선택된 이미지 파일들
+    const [imagePreviews, setImagePreviews] = useState([]); // 이미지 미리보기 URL들
+    const [showImageModal, setShowImageModal] = useState(false);
+    const [pptHistory, setPptHistory] = useState([]); // PPT 생성 기록 // 이미지 확대 모달
+    const [selectedImageForModal, setSelectedImageForModal] = useState(null); // 모달에 표시할 이미지
+    const [imageLoadingStates, setImageLoadingStates] = useState(new Map()); // 이미지 로딩 상태 추적
+    const [selectedExperience, setSelectedExperience] = useState(null); // 선택된 이력
+    const [showTemplateModal, setShowTemplateModal] = useState(false); // 템플릿 모달 표시 상태
+    const [selectedTemplateForModal, setSelectedTemplateForModal] = useState(null); // 모달에서 선택된 템플릿
+    const [showExperienceModal, setShowExperienceModal] = useState(false); // 이력 상세 모달 표시 여부
+    const [accessToken, setAccessToken] = useState('');
+    const [slides, setSlides] = useState([]);
+    const [presentationId, setPresentationId] = useState(null);
+    const [selectedTemplate, setSelectedTemplate] = useState('');
+    const [selectedExperiences, setSelectedExperiences] = useState([]);
+    const [driveViewMode, setDriveViewMode] = useState(() => {
+      // localStorage에서 저장된 뷰 모드 복원
+      return localStorage.getItem('driveViewMode') || 'all';
+    }); // 'all' 또는 'portfolio'
+    const [portfolioFolderId, setPortfolioFolderId] = useState(() => {
+      // localStorage에서 저장된 포트폴리오 폴더 ID 복원
+      return localStorage.getItem('portfolioFolderId') || null;
+    }); // 포트폴리오 폴더 ID
+    const formRef = useRef();
+  
+    // 통합 인증 서비스 인스턴스
+    const authService = useRef(new GoogleAuthService());
+    const sheetsService = useRef(null);
+    const driveService = useRef(null);
+  
+    // 섹션 전환
+    function showSection(section) {
+      setActiveSection(section);
+      localStorage.setItem('activeSection', section);
+    }
+  
+    // 통합 인증 시스템 초기화
+    async function initializeGoogleAuth() {
+  
+      try {
+        console.log('통합 인증 시스템 초기화 시작...');
+  
+        // 인증 상태 변경 리스너 등록
+        authService.current.addAuthStateListener((isAuthenticated) => {
+          setAuthStatus(isAuthenticated ? 'connected' : 'disconnected');
+          console.log('인증 상태 변경:', isAuthenticated);
+        });
+  
+        // 에러 리스너 등록
+        authService.current.addErrorListener((error) => {
+          console.error('인증 에러 발생:', error);
+          setAuthStatus('error');
+        });
+  
+        // 통합 인증 초기화
+        await authService.current.initialize();
+        console.log('통합 인증 시스템 초기화 완료');
+  
+        // 인증 상태 확인 및 토큰 갱신 시도
+        if (!authService.current.isAuthenticated()) {
+          console.log('인증 상태가 유효하지 않습니다. 토큰 갱신을 시도합니다...');
+          try {
+            // 토큰 갱신 시도 (팝업 없이)
+            await authService.current.refreshToken();
+            console.log('토큰 갱신 완료');
+          } catch (tokenError) {
+            console.log('토큰 갱신 실패:', tokenError);
+  
+            // interaction_required 오류는 정상적인 상황으로 처리
+            if (tokenError.message === 'interaction_required') {
+              console.log('사용자 상호작용이 필요한 상황입니다. 로그인 상태는 유지합니다.');
+  
+  
+              // 토큰 갱신이 실패해도 기존 토큰이 있다면 서비스 초기화를 시도
+              if (authService.current.hasExistingToken()) {
+                console.log('기존 토큰이 있습니다. 서비스 초기화를 시도합니다.');
+              } else {
+                console.log('기존 토큰이 없습니다. 서비스 초기화를 건너뜁니다.');
+                // 토큰이 없으면 서비스 초기화를 건너뜀 (로그인 상태는 유지)
+                setAuthStatus('disconnected');
+                setIsSheetsInitialized(false);
+                setIsDriveInitialized(false);
+                return;
+              }
+            } else {
+              // 다른 오류는 서비스 초기화를 건너뜀
+              setAuthStatus('disconnected');
+              setIsSheetsInitialized(false);
+              setIsDriveInitialized(false);
+              return;
+            }
+          }
+        }
+  
+        // 인증 완료 후 서비스들 초기화
+        await initializeServices();
+  
+      } catch (error) {
+        console.error('통합 인증 시스템 초기화 오류:', error);
+        setAuthStatus('error');
+        throw error;
+      }
+    }
+  
+    // 서비스들 초기화
+    async function initializeServices() {
+      try {
+        console.log('서비스들 초기화 시작...');
+  
+        // 인증 상태 확인
+        if (!authService.current.isAuthenticated()) {
+          console.log('인증이 완료되지 않았습니다. 서비스 초기화를 건너뜁니다.');
+          // 토큰이 없으면 서비스 초기화를 건너뜀
+          setAuthStatus('disconnected');
+          setIsSheetsInitialized(false);
+          setIsDriveInitialized(false);
+          return;
+        }
+  
+        // 서비스 인스턴스 생성 (의존성 주입)
+        sheetsService.current = new GoogleSheetsService(authService.current);
+        driveService.current = new GoogleDriveService(authService.current);
+  
+        console.log('서비스 인스턴스 생성 완료');
+  
+        // 기존 스프레드시트가 있는지 확인하고 없으면 생성
+        let currentSpreadsheetId = spreadsheetId;
+  
+        if (currentSpreadsheetId) {
+          console.log('기존 스프레드시트 ID 확인 중:', currentSpreadsheetId);
+  
+          try {
+            // 기존 시트가 실제로 존재하는지 확인
+            const exists = await sheetsService.current.checkSpreadsheetExists(currentSpreadsheetId);
+            if (!exists) {
+              console.log('기존 스프레드시트가 존재하지 않습니다. 상태를 초기화합니다...');
+              currentSpreadsheetId = null;
+              setSpreadsheetId(null);
+              localStorage.removeItem('spreadsheetId');
+            } else {
+              console.log('기존 스프레드시트가 유효합니다.');
+            }
+          } catch (error) {
+            console.log('기존 스프레드시트 확인 중 오류, 상태를 초기화합니다:', error);
+            currentSpreadsheetId = null;
+            setSpreadsheetId(null);
+            localStorage.removeItem('spreadsheetId');
+          }
+        }
+  
+        if (!currentSpreadsheetId) {
+          console.log('기존 포트폴리오 시트 파일 검색 중...');
+  
+          try {
+            // 포트폴리오 이력 폴더가 있는지 확인 (생성하지 않고 찾기만)
+            const portfolioFolder = await driveService.current.findFolder('포트폴리오 이력');
+  
+            if (portfolioFolder) {
+              console.log('기존 포트폴리오 이력 폴더 발견:', portfolioFolder.id);
+  
+              // 포트폴리오 이력 폴더 안에서 기존 시트 파일 검색
+              const existingFiles = await driveService.current.listFiles(50, portfolioFolder.id);
+              const portfolioFile = existingFiles.find(file =>
+                  file.name === '포트폴리오 이력' &&
+                  file.mimeType === 'application/vnd.google-apps.spreadsheet'
+              );
+  
+              if (portfolioFile) {
+                console.log('기존 포트폴리오 시트 파일 발견:', portfolioFile.id);
+                // 기존 파일 ID 저장
+                currentSpreadsheetId = portfolioFile.id;
+                setSpreadsheetId(currentSpreadsheetId);
+                localStorage.setItem('spreadsheetId', currentSpreadsheetId);
+  
+                // 포트폴리오 폴더 ID 설정
+                setPortfolioFolderId(portfolioFolder.id);
+                localStorage.setItem('portfolioFolderId', portfolioFolder.id);
+              } else {
+                console.log('포트폴리오 이력 폴더는 있지만 시트 파일이 없습니다.');
+                // 폴더는 있지만 시트가 없으면 폴더 ID만 저장
+                setPortfolioFolderId(portfolioFolder.id);
+                localStorage.setItem('portfolioFolderId', portfolioFolder.id);
+              }
+            } else {
+              console.log('포트폴리오 이력 폴더가 없습니다. 시트 생성 시 함께 생성됩니다.');
+            }
+  
+          } catch (error) {
+            console.error('기존 파일 확인 중 오류:', error);
+            console.log('시트 생성 시 폴더도 함께 생성됩니다.');
+          }
+        }
+  
+        // 서비스 초기화 상태 설정
+        setIsSheetsInitialized(true);
+        setIsDriveInitialized(true);
+  
+        // 기존 데이터 로드 (시트 생성 후에만 실행)
+        if (currentSpreadsheetId) {
+          // 시트 ID 상태를 먼저 업데이트
+          setSpreadsheetId(currentSpreadsheetId);
+  
+          // 새로 생성된 시트에서 데이터 로드
+          await loadExperiencesFromSheets(currentSpreadsheetId);
+          await loadDriveFiles();
+        }
+  
+        console.log('모든 서비스 초기화 완료');
+  
+      } catch (error) {
+        console.error('서비스 초기화 오류:', error);
+        const errorMessage = error?.message || '서비스 초기화에 실패했습니다.';
+        alert(errorMessage);
+        setIsSheetsInitialized(false);
+        setIsDriveInitialized(false);
+      }
+    }
+  
+    // 시트에서 이력 데이터 로드
+    async function loadExperiencesFromSheets(spreadsheetIdToUse = null) {
+      const targetSpreadsheetId = spreadsheetIdToUse || spreadsheetId;
+  
+      if (!targetSpreadsheetId || !sheetsService.current) return;
+  
+      try {
+        const sheetData = await sheetsService.current.readData(targetSpreadsheetId, 'A:E');
+        const experiences = sheetsService.current.formatSheetToExperience(sheetData);
+        setExperiences(experiences);
+        
+        // 이미지 프리로딩 (백그라운드에서 미리 로딩)
+        experiences.forEach(exp => {
+          if (exp.imageUrls && exp.imageUrls.length > 0) {
+            exp.imageUrls.forEach(imageUrl => {
+              preloadImage(imageUrl).catch(err => {
+                console.log('이미지 프리로딩 실패 (무시됨):', imageUrl, err);
+              });
+            });
+          }
+        });
+      } catch (error) {
+        console.error('이력 데이터 로드 오류:', error);
+        // 시트가 존재하지 않는 경우 로그만 출력하고 새로 생성하지 않음
+        if (error.message.includes('찾을 수 없습니다') || error.status === 404) {
+          console.log('시트가 존재하지 않습니다. 시트를 다시 생성해주세요.');
+          // 사용자에게 알림
+          alert('포트폴리오 시트가 삭제되었습니다. 로그아웃 후 다시 로그인해주세요.');
+        }
+      }
+    }
+  
+    // 드라이브 파일 목록 로드
+    async function loadDriveFiles(parentId = null) {
+      if (!driveService.current) return;
+  
+      try {
+        setIsDriveLoading(true);
+        console.log('드라이브 파일 불러오기 시작, 뷰 모드:', driveViewMode, '부모 ID:', parentId);
+  
+        // 시트가 있다면 실제로 존재하는지 확인
+        if (spreadsheetId && sheetsService.current) {
+          try {
+            const exists = await sheetsService.current.checkSpreadsheetExists(spreadsheetId);
+            if (!exists) {
+              console.log('저장된 시트가 존재하지 않습니다. 상태를 초기화합니다.');
+              setSpreadsheetId(null);
+              localStorage.removeItem('spreadsheetId');
+            }
+          } catch (error) {
+            console.log('시트 존재 확인 중 오류:', error);
+            setSpreadsheetId(null);
+            localStorage.removeItem('spreadsheetId');
+          }
+        }
+  
+        let files;
+        if (parentId) {
+          // 특정 폴더 내 파일 로드
+          files = await driveService.current.listFiles(50, parentId);
+          console.log('폴더 내 파일:', files);
+        } else if (driveViewMode === 'portfolio' && portfolioFolderId) {
+          // 포트폴리오 폴더 내 파일만 로드
+          files = await driveService.current.listFiles(50, portfolioFolderId);
+          console.log('포트폴리오 폴더 파일:', files);
+        } else {
+          // 전체 파일 로드
+          files = await driveService.current.listFiles(20);
+          console.log('전체 드라이브 파일:', files);
+        }
+  
+        setDriveFiles(files);
+      } catch (error) {
+        console.error('드라이브 파일 로드 오류:', error);
+      } finally {
+        setIsDriveLoading(false);
+      }
+    }
+  
+    // 로그인 상태 저장
+    function saveLoginState(loggedIn, spreadsheetIdValue = null) {
+      setIsLoggedIn(loggedIn);
+      localStorage.setItem('isLoggedIn', loggedIn.toString());
+  
+      if (spreadsheetIdValue) {
+        setSpreadsheetId(spreadsheetIdValue);
+        localStorage.setItem('spreadsheetId', spreadsheetIdValue);
+      }
+  
+      // 로그아웃 시에는 스프레드시트 ID도 제거
+      if (!loggedIn) {
+        localStorage.removeItem('spreadsheetId');
+        setSpreadsheetId(null);
+      }
+    }
+  
+    // GIS 기반 로그인 (단일 팝업에서 로그인+권한 처리)
+    async function handleGISLogin() {
+      try {
+        setIsLoading(true);
+        console.log('GIS 기반 로그인 시작...');
+  
+        // 통합 인증 시스템 초기화
+        await authService.current.initialize();
+        console.log('인증 시스템 초기화 완료');
+  
+        // 단일 팝업에서 로그인과 권한 요청
+        await authService.current.requestToken();
+        console.log('GIS 로그인 및 권한 요청 완료');
+  
+        // 로그인 상태 저장 (이미 isLoggedIn이 true로 설정되어 있으므로 스프레드시트 ID만 전달)
+        saveLoginState(true);
+  
+        // 메인 페이지로 이동 (저장된 섹션이 있으면 그대로 유지)
+        const savedSection = localStorage.getItem('activeSection') || 'main';
+        setActiveSection(savedSection);
+  
+        // 인증 완료 후 서비스들 초기화
+        await initializeServices();
+  
+        // 액세스 토큰 설정
+        try {
+          const token = authService.current.getAccessToken();
+          setAccessToken(token);
+        } catch (error) {
+          console.log('토큰 가져오기 실패:', error);
+        }
+  
+      } catch (error) {
+        console.error('GIS 로그인 오류:', error);
+        const errorMessage = error?.message || '로그인에 실패했습니다.';
+        alert(errorMessage);
+      } finally {
+        setIsLoading(false);
+      }
+    }
+  
+    // 로그아웃
+    function logout() {
+      if (window.confirm('로그아웃 하시겠습니까?')) {
+        // 통합 인증 서비스에서 로그아웃
+        authService.current.logout();
+  
+        // 로컬 상태 정리 (saveLoginState에서 스프레드시트 ID도 제거됨)
+        saveLoginState(false);
+        setActiveSection('main');
+        localStorage.setItem('activeSection', 'main');
+        localStorage.setItem('driveViewMode', 'all');
+        localStorage.removeItem('portfolioFolderId');
+        setPortfolioFolderId(null);
+        setIsSheetsInitialized(false);
+        setIsDriveInitialized(false);
+        setAuthStatus('disconnected');
+  
+        // 서비스 인스턴스 정리
+        sheetsService.current = null;
+        driveService.current = null;
+      }
+    }
+  
+    function showEditExperienceModal(index) {
+      if (index === null || index < 0 || index >= experiences.length) {
+        console.error('유효하지 않은 이력 인덱스:', index);
+        alert('수정할 이력이 존재하지 않습니다.');
+        return;
+      }
+  
+      const expToEdit = experiences[index];
+  
+      // 1. period 필드에서 날짜를 파싱합니다
+      let startDate = '';
+      let endDate = '';
+      
+      // 디버깅을 위해 데이터 구조 확인
+      console.log('수정할 이력 데이터:', expToEdit);
+      
+      if (expToEdit.period) {
+        console.log('period 필드:', expToEdit.period);
+        
+        // 다양한 period 형식 지원
+        // 1. "2025.06.30 - 2025.10.30" 형식
+        let periodMatch = expToEdit.period.match(/(\d{4}\.\d{2}\.\d{2})\s*[-~]\s*(\d{4}\.\d{2}\.\d{2})/);
+        
+        if (periodMatch) {
+          // "2025.06.30" 형식을 "2025-06-30" 형식으로 변환
+          startDate = periodMatch[1].replace(/\./g, '-');
+          endDate = periodMatch[2].replace(/\./g, '-');
+          console.log('파싱된 날짜:', { startDate, endDate });
+        } else {
+          // 2. "2025-06-30 - 2025-10-30" 형식 (이미 하이픈 형식)
+          periodMatch = expToEdit.period.match(/(\d{4}-\d{2}-\d{2})\s*[-~]\s*(\d{4}-\d{2}-\d{2})/);
+          if (periodMatch) {
+            startDate = periodMatch[1];
+            endDate = periodMatch[2];
+            console.log('파싱된 날짜 (하이픈 형식):', { startDate, endDate });
+          } else {
+            // 3. "2025/06/30 - 2025/10/30" 형식
+            periodMatch = expToEdit.period.match(/(\d{4}\/\d{2}\/\d{2})\s*[-~]\s*(\d{4}\/\d{2}\/\d{2})/);
+            if (periodMatch) {
+              startDate = periodMatch[1].replace(/\//g, '-');
+              endDate = periodMatch[2].replace(/\//g, '-');
+              console.log('파싱된 날짜 (슬래시 형식):', { startDate, endDate });
+            } else {
+              // 4. 단일 날짜 형식 "2025.09.24" (시작일만 있는 경우)
+              periodMatch = expToEdit.period.match(/(\d{4}\.\d{2}\.\d{2})/);
+              if (periodMatch) {
+                startDate = periodMatch[1].replace(/\./g, '-');
+                endDate = periodMatch[1].replace(/\./g, '-'); // 시작일과 동일하게 설정
+                console.log('파싱된 날짜 (단일 날짜):', { startDate, endDate });
+              } else {
+                // 5. 하이픈 형식 단일 날짜 "2025-09-24"
+                periodMatch = expToEdit.period.match(/(\d{4}-\d{2}-\d{2})/);
+                if (periodMatch) {
+                  startDate = periodMatch[1];
+                  endDate = periodMatch[1]; // 시작일과 동일하게 설정
+                  console.log('파싱된 날짜 (하이픈 단일 날짜):', { startDate, endDate });
+                } else {
+                  console.log('날짜 파싱 실패 - 지원되지 않는 형식:', expToEdit.period);
+                }
+              }
+            }
+          }
+        }
+      } else {
+        console.log('period 필드가 없습니다');
+      }
+  
+      // 2. form 상태 초기화 (날짜 입력 필드의 value 역할)
+      setForm({
+        title: expToEdit.title || '',
+        startDate: startDate, // ⭐ 기존 데이터를 사용하여 달력 필드에 표시
+        endDate: endDate,     // ⭐ 기존 데이터를 사용하여 달력 필드에 표시
+        description: expToEdit.description || ''
+      });
+  
+      // ⭐ [핵심] 기존 설정 기간을 저장합니다. ⭐
+      //    모달이 열리자마자 이 값이 '기존 설정 기간' 텍스트에 사용됩니다.
+      setOriginalPeriod({
+        start: startDate, // "2025.06.30"
+        end: endDate      // "2025.10.30"
+      });
+  
+      // 3. 이미지 및 모드 설정
+      const existingUrls = expToEdit.imageUrls || [];
+      setImagePreviews(existingUrls);
+  
+      setEditingIndex(index);
+      setShowModal(true);
+    }
+  
+    function showAddExperienceModal() {
+      console.log('showAddExperienceModal 호출됨');
+      // 상태 초기화 (폼, 이미지, 편집 인덱스 등)
+      setForm({ title: '', startDate: '', endDate: '', description: '' });
+      setSelectedImages([]);
+      setImagePreviews([]);
+      setEditingIndex(null); // 수정 모드 아님을 확실히 지정
+  
+      setShowModal(true);
+      console.log('showModal 상태를 true로 설정');
+    }
+    function closeModal() {
+      setShowModal(false);
+      setForm({ title: '', startDate: '', endDate: '', description: '' });
+      setSelectedImages([]);
+      setImagePreviews([]);
+      setEditingIndex(null);
+  
+      setOriginalPeriod(null);
+    }
+  
+    // 기간 포맷팅 함수
+    function formatPeriod(startDate, endDate) {
+      if (!startDate || !endDate) return '';
+  
+      const start = new Date(startDate);
+      const end = new Date(endDate);
+  
+      // 시작일과 종료일이 같은 경우 (하루짜리)
+      if (start.toDateString() === end.toDateString()) {
+        const year = start.getFullYear();
+        const month = String(start.getMonth() + 1).padStart(2, '0');
+        const day = String(start.getDate()).padStart(2, '0');
+        return `${year}.${month}.${day}`;
+      }
+  
+      // 여러 기간인 경우
+      const startYear = start.getFullYear();
+      const startMonth = String(start.getMonth() + 1).padStart(2, '0');
+      const startDay = String(start.getDate()).padStart(2, '0');
+      const endYear = end.getFullYear();
+      const endMonth = String(end.getMonth() + 1).padStart(2, '0');
+      const endDay = String(end.getDate()).padStart(2, '0');
+  
+      return `${startYear}.${startMonth}.${startDay} - ${endYear}.${endMonth}.${endDay}`;
+    }
+  
+    // 날짜 유효성 검사 함수
+    function validateDates(startDate, endDate) {
+      if (!startDate || !endDate) return false;
+  
+      const start = new Date(startDate);
+      const end = new Date(endDate);
+  
+      // 시작일이 종료일보다 늦으면 안됨
+      if (start > end) {
+        alert('시작일은 종료일보다 이전이어야 합니다.');
+        return false;
+      }
+  
+      // 미래 날짜 체크 제거 (너무 엄격함)
+      // const today = new Date();
+      // if (start > today) {
+      //   alert('시작일은 오늘 이전이어야 합니다.');
+      //   return false;
+      // }
+  
+      return true;
+    }
+  
+    // 이력 저장
+    async function saveExperience(e) {
+      e.preventDefault();
+      if (form.title && form.startDate && form.endDate && form.description) {
+        // 날짜 유효성 검사
+        if (!validateDates(form.startDate, form.endDate)) {
+          return;
+        }
+  
+        try {
+          setIsExperienceLoading(true);
+  
+          let newImageUrls = [];
+          // 1. 새 이미지가 있다면 드라이브에 업로드
+          if (selectedImages.length > 0) {
+            for (const imageFile of selectedImages) {
+              const imageUrl = await uploadImageToDrive(imageFile, form.title);
+              newImageUrls.push(imageUrl);
+            }
+          }
+  
+          const period = formatPeriod(form.startDate, form.endDate);
+          let experienceToSave;
+  
+          if (editingIndex !== null) {
+            // ⭐ 수정 모드 (Editing Mode)
+            const existingImageUrls = experiences[editingIndex].imageUrls || [];
+  
+            experienceToSave = {
+              ...form,
+              period,
+              // 기존 이미지 URL에 새로 업로드한 이미지 URL을 합칩니다.
+              imageUrls: existingImageUrls.concat(newImageUrls)
+            };
+  
+            const rowNumber = editingIndex + 2;
+            if (spreadsheetId && sheetsService.current) {
+              const sheetData = sheetsService.current.formatExperienceForSheet(experienceToSave);
+              await sheetsService.current.updateData(spreadsheetId, `A${rowNumber}:E${rowNumber}`, [sheetData]);
+            }
+            setExperiences(prev => prev.map((exp, i) => i === editingIndex ? experienceToSave : exp));
+  
+          } else {
+            // ⭐ 등록 모드 (Creation Mode)
+            experienceToSave = {
+              ...form,
+              period,
+              imageUrls: newImageUrls // 새 이력은 새로 업로드한 이미지 URL만 가집니다.
+            };
+  
+            // 1. 구글 시트 추가
+            if (spreadsheetId && sheetsService.current) {
+              const sheetData = sheetsService.current.formatExperienceForSheet(experienceToSave);
+              await sheetsService.current.appendData(spreadsheetId, 'A:E', [sheetData]);
+            }
+  
+            // 2. ⭐ 로컬 상태에 새 이력 추가 (정확히 새 이력 객체만 추가) ⭐
+            // 이 코드가 이력을 성공적으로 추가하고 화면에 남아있도록 보장합니다.
+            setExperiences(prev => [...prev, experienceToSave]);
+          }
+  
+          closeModal();
+        } catch (error) {
+          console.error('이력 저장/수정 오류:', error);
+          const errorMessage = sheetsService.current?.formatErrorMessage(error) || '이력 저장/수정에 실패했습니다.';
+          alert(errorMessage);
+        } finally {
+          setIsExperienceLoading(false);
+        }
+      }
+    }
+  
+    // 전체 선택/해제
+    function selectAllExperiences(select) {
+      if (select) {
+        setSelected(experiences.map((_, i) => i));
+      } else {
+        setSelected([]);
+      }
+    }
+  
+    // 체크박스 변경
+    function toggleSelect(idx) {
+      setSelected(selected.includes(idx)
+          ? selected.filter(i => i !== idx)
+          : [...selected, idx]
+      );
+    }
+  
+    // 이미지 선택 핸들러
+    function handleImageSelect(event) {
+      const files = Array.from(event.target.files);
+  
+      // 각 파일에 대해 검증 및 처리
+      files.forEach(file => {
+        // 파일 크기 체크 (5MB 제한)
+        if (file.size > 5 * 1024 * 1024) {
+          alert(`파일 ${file.name}의 크기는 5MB 이하여야 합니다.`);
+          return;
+        }
+  
+        // 이미지 파일 타입 체크
+        if (!file.type.startsWith('image/')) {
+          alert(`파일 ${file.name}은 이미지 파일이 아닙니다.`);
+          return;
+        }
+  
+        // 이미지 미리보기 URL 생성
+        const reader = new FileReader();
+        reader.onload = (e) => {
+          setSelectedImages(prev => [...prev, file]);
+          setImagePreviews(prev => [...prev, e.target.result]);
+        };
+        reader.readAsDataURL(file);
+      });
+  
+      // input 초기화 (같은 파일을 다시 선택할 수 있도록)
+      event.target.value = '';
+    }
+  
+    // 이미지 제거
+    function removeImage(index) {
+      setSelectedImages(prev => prev.filter((_, i) => i !== index));
+      setImagePreviews(prev => prev.filter((_, i) => i !== index));
+    }
+  
+    // 이미지 확대 모달 표시
+    function openImageModal(imageUrl, title) {
+      setSelectedImageForModal({ url: imageUrl, title });
+      setShowImageModal(true);
+    }
+  
+    // 이력 상세 모달 열기
+    function openExperienceModal(experience) {
+      setSelectedExperience(experience);
+      setShowExperienceModal(true);
+    }
+  
+    // 이력 상세 모달 닫기
+    function closeExperienceModal() {
+      setSelectedExperience(null);
+      setShowExperienceModal(false);
+    }
+  
+    // 이미지 확대 모달 닫기
+    function closeImageModal() {
+      setShowImageModal(false);
+      setSelectedImageForModal(null);
+    }
+  
+    // 이미지 URL 변환 캐시 (무한 재귀 방지)
+    const imageUrlCache = new Map();
+    
+    // 이미지 프리로딩 캐시 (이미 로딩된 이미지 추적)
+    const preloadedImages = new Set();
+    
+    // 이미지 프리로딩 함수 (백그라운드에서 미리 로딩)
+    function preloadImage(imageUrl) {
+      if (preloadedImages.has(imageUrl)) {
+        return Promise.resolve();
+      }
+      
+      return new Promise((resolve, reject) => {
+        const img = new Image();
+        img.onload = () => {
+          preloadedImages.add(imageUrl);
+          resolve();
+        };
+        img.onerror = reject;
+        img.src = imageUrl;
+      });
+    }
+    
+    // 이미지 로딩 상태 관리 함수들
+    function setImageLoadingState(imageKey, isLoading) {
+      setImageLoadingStates(prev => {
+        const newMap = new Map(prev);
+        if (isLoading) {
+          newMap.set(imageKey, 'loading');
+        } else {
+          newMap.delete(imageKey);
+        }
+        return newMap;
+      });
+    }
+  
+    function setImageErrorState(imageKey) {
+      setImageLoadingStates(prev => {
+        const newMap = new Map(prev);
+        newMap.set(imageKey, 'error');
+        return newMap;
+      });
+    }
+  
+    // 이미지 로딩 재시도 함수 (최적화됨)
+    async function retryImageLoad(imgElement, originalUrl, retryCount = 0) {
+      const maxRetries = 2; // 재시도 횟수 감소
+      const imageKey = `${originalUrl}_${imgElement.alt}`;
+      
+      // 이미 프리로딩된 이미지인지 확인
+      if (preloadedImages.has(originalUrl)) {
+        imgElement.src = originalUrl;
+        return;
+      }
+      
+      const fileId = originalUrl.match(/[-\w]{25,}/)?.[0];
+      const alternativeUrls = [
+        `https://drive.google.com/thumbnail?id=${fileId}&sz=w400`, // 가장 빠른 썸네일
+        `https://drive.google.com/thumbnail?id=${fileId}&sz=w200`, // 더 작은 썸네일
+        `https://drive.google.com/uc?export=view&id=${fileId}`
+      ];
+      
+      if (retryCount >= maxRetries) {
+        console.error('이미지 로딩 실패 - 모든 재시도 소진:', originalUrl);
+        setImageErrorState(imageKey);
+        imgElement.style.display = 'none';
+        return;
+      }
+      
+      const currentUrl = retryCount === 0 ? originalUrl : alternativeUrls[retryCount - 1];
+      console.log(`이미지 로딩 재시도 ${retryCount + 1}/${maxRetries + 1}:`, currentUrl);
+      
+      // 로딩 상태 설정
+      setImageLoadingState(imageKey, true);
+      
+      // 새로운 이미지 객체로 테스트 (타임아웃 설정)
+      const testImg = new Image();
+      let timeoutId;
+      
+      const cleanup = () => {
+        if (timeoutId) clearTimeout(timeoutId);
+        setImageLoadingState(imageKey, false);
+      };
+      
+      // 3초 타임아웃 설정 (더 빠른 응답)
+      timeoutId = setTimeout(() => {
+        console.log('이미지 로딩 타임아웃:', currentUrl);
+        cleanup();
+        setTimeout(() => retryImageLoad(imgElement, originalUrl, retryCount + 1), 300); // 더 빠른 재시도
+      }, 3000);
+      
+      testImg.onload = () => {
+        cleanup();
+        preloadedImages.add(currentUrl); // 성공한 URL을 캐시에 추가
+        imgElement.src = currentUrl;
+        console.log('이미지 로딩 성공:', currentUrl);
+      };
+      
+      testImg.onerror = () => {
+        cleanup();
+        setTimeout(() => retryImageLoad(imgElement, originalUrl, retryCount + 1), 300); // 더 빠른 재시도
+      };
+      
+      testImg.src = currentUrl;
+    }
+  
+    // 기존 이미지 URL을 올바른 형식으로 변환
+    async function convertImageUrl(imageUrl) {
+      // 캐시에서 확인
+      if (imageUrlCache.has(imageUrl)) {
+        return imageUrlCache.get(imageUrl);
+      }
+  
+      // 이미 Base64 데이터 URL이거나 직접 접근 URL인 경우 그대로 반환
+      if (imageUrl.startsWith('data:') || imageUrl.includes('uc?export=view&id=')) {
+        imageUrlCache.set(imageUrl, imageUrl);
+        return imageUrl;
+      }
+  
+      // 구글 드라이브 파일 ID 추출
+      const fileIdMatch = imageUrl.match(/[-\w]{25,}/);
+      if (!fileIdMatch) {
+        console.warn('구글 드라이브 파일 ID를 찾을 수 없습니다:', imageUrl);
+        imageUrlCache.set(imageUrl, imageUrl);
+        return imageUrl;
+      }
+  
+      const fileId = fileIdMatch[0];
+      
+      // 썸네일 크기로 최적화된 URL 우선 사용 (로딩 속도 향상)
+      const possibleUrls = [
+        `https://drive.google.com/thumbnail?id=${fileId}&sz=w400`, // 썸네일 크기로 최적화
+        `https://drive.google.com/uc?export=view&id=${fileId}`,
+        `https://drive.google.com/thumbnail?id=${fileId}&sz=w1000`,
+        `https://lh3.googleusercontent.com/d/${fileId}`,
+        `https://drive.google.com/file/d/${fileId}/view?usp=sharing`
+      ];
+      
+      // 첫 번째 URL을 기본으로 사용 (썸네일 크기로 빠른 로딩)
+      const directUrl = possibleUrls[0];
+      console.log('이미지 URL 변환 (최적화):', imageUrl, '→', directUrl);
+      
+      imageUrlCache.set(imageUrl, directUrl);
+      return directUrl;
+    }
+  
+    // 구글 계정 이름 가져오기
+    async function getGoogleAccountName() {
+      try {
+        const gapiClient = authService.current.getAuthenticatedGapiClient();
+        
+        // Google People API를 사용하여 사용자 정보 가져오기
+        const response = await gapiClient.people.people.get({
+          resourceName: 'people/me',
+          personFields: 'names'
+        });
+        
+        const names = response.result.names;
+        if (names && names.length > 0) {
+          // 첫 번째 이름 반환 (일반적으로 주 이름)
+          return names[0].displayName || names[0].givenName || '사용자';
+        }
+        
+        return '사용자';
+      } catch (error) {
+        console.error('구글 계정 이름 가져오기 실패:', error);
+        return '사용자';
+      }
+    }
+  
+    // PPT 기록 조회
+    async function loadPptHistory() {
+      if (!driveService.current) return;
+      
+      try {
+        setIsLoading(true);
+        
+        // 포트폴리오 폴더 확인/생성
+        const portfolioFolder = await driveService.current.ensurePortfolioFolder();
+        
+        // PPT 폴더 찾기
+        const pptFolder = await driveService.current.findFolder('PPT', portfolioFolder.id);
+        
+        if (!pptFolder) {
+          console.log('PPT 폴더가 없습니다.');
+          setPptHistory([]);
+          return;
+        }
+        
+        // PPT 폴더에서 PPT 파일들 조회
+        const files = await driveService.current.getFilesInFolder(pptFolder.id);
+        
+        // PPT 파일만 필터링 (Google Slides 파일)
+        const pptFiles = files.filter(file => 
+          file.mimeType === 'application/vnd.google-apps.presentation'
+        );
+        
+        // 생성일 기준으로 최신순 정렬
+        const sortedPptFiles = pptFiles.sort((a, b) => 
+          new Date(b.createdTime) - new Date(a.createdTime)
+        );
+        
+        setPptHistory(sortedPptFiles);
+        console.log('PPT 기록 로드됨:', sortedPptFiles);
+        
+      } catch (error) {
+        console.error('PPT 기록 조회 오류:', error);
+        alert('PPT 기록을 불러오는데 실패했습니다.');
+      } finally {
+        setIsLoading(false);
+      }
+    }
+  
+    // PPT 수정을 위한 슬라이드 데이터 로드
+    async function loadPptForEdit(pptId) {
+      try {
+        setIsLoading(true);
+        console.log('PPT 수정을 위한 데이터 로드 시작:', pptId);
+        
+        // 인증 서비스를 통해 토큰 확인 및 갱신
+        if (!authService.current) {
+          alert('인증 서비스가 초기화되지 않았습니다. 다시 로그인해주세요.');
+          return;
+        }
+        
+        // 인증 상태 확인
+        const isAuthenticated = await authService.current.isAuthenticated();
+        if (!isAuthenticated) {
+          alert('인증이 필요합니다. 다시 로그인해주세요.');
+          return;
+        }
+        
+        // 액세스 토큰 가져오기
+        const token = await authService.current.getAccessToken();
+        if (!token) {
+          alert('액세스 토큰을 가져올 수 없습니다. 다시 로그인해주세요.');
+          return;
+        }
+        
+        // 프레젠테이션 데이터 가져오기
+        const data = await getPresentationData(pptId, token);
+        console.log('프레젠테이션 데이터 로드됨:', data);
+        
+        // 슬라이드 데이터 설정
+        setSlides(data.slides || []);
+        setPresentationId(pptId);
+        
+        console.log('슬라이드 데이터 설정 완료:', data.slides);
+        
+        // 에디터 섹션으로 이동
+        setActiveSection('editor');
+        
+      } catch (error) {
+        console.error('PPT 수정 데이터 로드 오류:', error);
+        if (error.message && error.message.includes('인증')) {
+          alert('인증이 만료되었습니다. 다시 로그인해주세요.');
+        } else {
+          alert('PPT 데이터를 불러오는데 실패했습니다: ' + (error.message || error));
+        }
+      } finally {
+        setIsLoading(false);
+      }
+    }
+  
+    // 드롭된 파일 처리
+    function handleDroppedFiles(files) {
+      files.forEach(file => {
+        // 파일 크기 체크 (5MB 제한)
+        if (file.size > 5 * 1024 * 1024) {
+          alert(`파일 ${file.name}의 크기는 5MB 이하여야 합니다.`);
+          return;
+        }
+  
+        // 이미지 파일 타입 체크
+        if (!file.type.startsWith('image/')) {
+          alert(`파일 ${file.name}은 이미지 파일이 아닙니다.`);
+          return;
+        }
+  
+        // 이미지 미리보기 URL 생성
+        const reader = new FileReader();
+        reader.onload = (e) => {
+          setSelectedImages(prev => [...prev, file]);
+          setImagePreviews(prev => [...prev, e.target.result]);
+        };
+        reader.readAsDataURL(file);
+      });
+    }
+  
+    // 이미지를 구글 드라이브에 업로드하고 공개 링크 생성
+    async function uploadImageToDrive(imageFile, experienceTitle) {
+      if (!driveService.current) {
+        throw new Error('구글 드라이브 서비스가 초기화되지 않았습니다.');
+      }
+  
+      try {
+        // 포트폴리오 이력 폴더와 이미지 폴더 확인/생성
+        const portfolioFolder = await driveService.current.ensurePortfolioFolder();
+        const imageFolder = await driveService.current.ensureImageFolder(portfolioFolder.id);
+  
+        // 이력별 이미지 폴더 생성 또는 찾기
+        const experienceFolder = await driveService.current.ensureExperienceImageFolder(experienceTitle, imageFolder.id);
+  
+        // 해당 이력 폴더의 기존 파일 목록 가져오기
+        const existingFiles = await driveService.current.getFilesInFolder(experienceFolder.id);
+        
+        // 순차적 번호가 포함된 파일명 생성
+        const finalFileName = driveService.current.generateSequentialFileName(imageFile.name, existingFiles);
+        console.log(`원본 파일명: ${imageFile.name} → 최종 파일명: ${finalFileName}`);
+  
+        // 이미지 파일을 해당 이력 폴더에 업로드
+        const uploadResult = await driveService.current.uploadFile(
+            finalFileName,
+            imageFile,
+            imageFile.type,
+            experienceFolder.id
+        );
+  
+        if (!uploadResult.id) {
+          throw new Error('이미지 업로드에 실패했습니다.');
+        }
+  
+        // 파일을 공개로 설정 (링크 공유 가능하게)
+        const gapiClient = authService.current.getAuthenticatedGapiClient();
+        await gapiClient.drive.permissions.create({
+          fileId: uploadResult.id,
+          resource: {
+            role: 'reader',
+            type: 'anyone'
+          }
+        });
+  
+        // 권한 설정 후 충분한 대기 시간 (권한 변경이 반영되도록)
+        await new Promise(resolve => setTimeout(resolve, 3000));
+  
+        // 썸네일 크기로 최적화된 URL 반환 (빠른 로딩)
+        const directUrl = `https://drive.google.com/thumbnail?id=${uploadResult.id}&sz=w400`;
+        console.log('생성된 이미지 URL (최적화):', directUrl);
+        return directUrl;
+      } catch (error) {
+        console.error('이미지 업로드 오류:', error);
+        throw new Error('이미지를 구글 드라이브에 업로드하는데 실패했습니다.');
+      }
+    }
+  
+    // 이미지 URL에서 파일 ID 추출하는 헬퍼 함수
+    function extractFileIdFromImageUrl(imageUrl) {
+      try {
+        // Google Drive 썸네일 URL에서 파일 ID 추출
+        // 예: https://drive.google.com/thumbnail?id=FILE_ID&sz=w400
+        const match = imageUrl.match(/[?&]id=([^&]+)/);
+        if (match) {
+          return match[1];
+        }
+        
+        // 직접 파일 URL에서 파일 ID 추출
+        // 예: https://drive.google.com/file/d/FILE_ID/view
+        const directMatch = imageUrl.match(/\/file\/d\/([^\/]+)/);
+        if (directMatch) {
+          return directMatch[1];
+        }
+        
+        return null;
+      } catch (error) {
+        console.error('파일 ID 추출 오류:', error);
+        return null;
+      }
+    }
+  
+    // 이미지 파일들을 삭제하는 헬퍼 함수
+    async function deleteImageFiles(imageUrls) {
+      if (!imageUrls || imageUrls.length === 0 || !driveService.current) return;
+      
+      const deletePromises = imageUrls.map(async (imageUrl) => {
+        try {
+          const fileId = extractFileIdFromImageUrl(imageUrl);
+          if (fileId) {
+            console.log('이미지 파일 삭제 시도:', fileId);
+            await driveService.current.deleteFile(fileId);
+            console.log('이미지 파일 삭제 완료:', fileId);
+          }
+        } catch (error) {
+          console.error('이미지 파일 삭제 실패:', imageUrl, error);
+          // 개별 이미지 삭제 실패해도 전체 프로세스는 계속 진행
+        }
+      });
+      
+      await Promise.allSettled(deletePromises);
+    }
+  
+    // 이력 폴더를 삭제하는 헬퍼 함수
+    async function deleteExperienceFolder(experienceTitle) {
+      if (!experienceTitle || !driveService.current || !portfolioFolderId) {
+        console.log('폴더 삭제 조건 미충족:', { experienceTitle, hasDriveService: !!driveService.current, portfolioFolderId });
+        return;
+      }
+      
+      try {
+        console.log('포트폴리오 폴더 ID:', portfolioFolderId);
+        console.log('검색할 이력 제목:', experienceTitle);
+        
+        // 포트폴리오 이력 폴더 내에서 image 폴더 찾기
+        const files = await driveService.current.getFilesInFolder(portfolioFolderId);
+        console.log('포트폴리오 폴더 내 파일들:', files.map(f => ({ name: f.name, mimeType: f.mimeType, id: f.id })));
+        
+        const imageFolder = files.find(file => 
+          file.name === 'image' && 
+          file.mimeType === 'application/vnd.google-apps.folder'
+        );
+        
+        if (!imageFolder) {
+          console.log('image 폴더를 찾을 수 없습니다');
+          return;
+        }
+        
+        console.log('image 폴더 발견:', imageFolder.name, imageFolder.id);
+        
+        // image 폴더 내에서 해당 이력 폴더 찾기
+        const imageFiles = await driveService.current.getFilesInFolder(imageFolder.id);
+        console.log('image 폴더 내 파일들:', imageFiles.map(f => ({ name: f.name, mimeType: f.mimeType, id: f.id })));
+        
+        // 정확한 이름으로 먼저 찾기
+        let experienceFolder = imageFiles.find(file => 
+          file.name === experienceTitle && 
+          file.mimeType === 'application/vnd.google-apps.folder'
+        );
+        
+        // 정확한 이름으로 찾지 못했다면 부분 일치로 찾기
+        if (!experienceFolder) {
+          experienceFolder = imageFiles.find(file => 
+            file.name.includes(experienceTitle) && 
+            file.mimeType === 'application/vnd.google-apps.folder'
+          );
+        }
+        
+        // 여전히 찾지 못했다면 대소문자 무시하고 찾기
+        if (!experienceFolder) {
+          experienceFolder = imageFiles.find(file => 
+            file.name.toLowerCase() === experienceTitle.toLowerCase() && 
+            file.mimeType === 'application/vnd.google-apps.folder'
+          );
+        }
+        
+        if (experienceFolder) {
+          console.log('이력 폴더 발견:', experienceFolder.name, experienceFolder.id);
+          console.log('이력 폴더 삭제 시도:', experienceFolder.name, experienceFolder.id);
+          await driveService.current.deleteFile(experienceFolder.id);
+          console.log('이력 폴더 삭제 완료:', experienceFolder.name);
+        } else {
+          console.log('삭제할 이력 폴더를 찾을 수 없습니다:', experienceTitle);
+          console.log('image 폴더 내 사용 가능한 폴더들:', imageFiles.filter(f => f.mimeType === 'application/vnd.google-apps.folder').map(f => f.name));
+          console.log('검색 조건:', { 
+            exactMatch: experienceTitle,
+            partialMatch: `contains ${experienceTitle}`,
+            caseInsensitive: experienceTitle.toLowerCase()
+          });
+        }
+      } catch (error) {
+        console.error('이력 폴더 삭제 실패:', experienceTitle, error);
+        // 폴더 삭제 실패해도 전체 프로세스는 계속 진행
+      }
+    }
+  
+    // 선택된 이력 삭제
+    async function deleteSelectedExperiences() {
+      if (selected.length === 0 || !sheetsService.current) return;
+  
+      if (!window.confirm('선택된 이력을 삭제하시겠습니까?')) return;
+  
+      try {
+        setIsExperienceLoading(true);
+  
+        // 삭제할 이력들의 이미지 URL과 제목 수집
+        const imagesToDelete = [];
+        const foldersToDelete = [];
+        for (const index of selected) {
+          const exp = experiences[index];
+          if (exp.imageUrls && exp.imageUrls.length > 0) {
+            imagesToDelete.push(...exp.imageUrls);
+          }
+          if (exp.title) {
+            foldersToDelete.push(exp.title);
+          }
+        }
+  
+        // 선택된 이력들을 구글 시트에서 삭제
+        if (spreadsheetId && isSheetsInitialized) {
+          // 선택된 행들을 역순으로 정렬하여 삭제 (인덱스가 변경되지 않도록)
+          const sortedSelected = [...selected].sort((a, b) => b - a);
+  
+          for (const index of sortedSelected) {
+            // 헤더 + 선택된 인덱스 + 1 (시트는 1부터 시작)
+            const rowNumber = index + 2;
+            await sheetsService.current.deleteData(spreadsheetId, `A${rowNumber}:E${rowNumber}`);
+          }
+        }
+  
+        // 이미지 파일들 삭제
+        if (imagesToDelete.length > 0) {
+          console.log('삭제할 이미지 파일들:', imagesToDelete);
+          await deleteImageFiles(imagesToDelete);
+        }
+  
+        // 이력 폴더들 삭제
+        if (foldersToDelete.length > 0) {
+          console.log('삭제할 이력 폴더들:', foldersToDelete);
+          const folderDeletePromises = foldersToDelete.map(folderName => 
+            deleteExperienceFolder(folderName)
+          );
+          await Promise.allSettled(folderDeletePromises);
+        }
+  
+        // 로컬 상태에서도 삭제
+        const newExperiences = experiences.filter((_, idx) => !selected.includes(idx));
+        setExperiences(newExperiences);
+        setSelected([]);
+  
+      } catch (error) {
+        console.error('이력 삭제 오류:', error);
+        const errorMessage = sheetsService.current?.formatErrorMessage(error) || '이력 삭제에 실패했습니다.';
+        alert(errorMessage);
+      } finally {
+        setIsExperienceLoading(false);
+      }
+    }
+  
+    // 개별 이력 삭제
+    async function deleteIndividualExperience(indexToDelete) {
+      const expToDelete = experiences[indexToDelete];
+  
+      if (!window.confirm(`"${expToDelete.title}" 이력을 정말로 삭제하시겠습니까?`)) {
+        return;
+      }
+  
+      try {
+        setIsExperienceLoading(true);
+  
+        // 삭제할 이력의 이미지 URL 수집
+        const imagesToDelete = expToDelete.imageUrls || [];
+  
+        if (spreadsheetId && sheetsService.current) {
+          // 헤더(1행) + 인덱스(0부터 시작) + 1 = 실제 시트 행 번호
+          const rowNumber = indexToDelete + 2;
+  
+          // 시트에서 해당 행 삭제
+          await sheetsService.current.deleteData(spreadsheetId, `A${rowNumber}:E${rowNumber}`);
+        }
+  
+        // 이미지 파일들 삭제
+        if (imagesToDelete.length > 0) {
+          console.log('삭제할 이미지 파일들:', imagesToDelete);
+          await deleteImageFiles(imagesToDelete);
+        }
+  
+        // 이력 폴더 삭제
+        if (expToDelete.title) {
+          console.log('개별 이력 삭제 - 삭제할 이력 폴더:', expToDelete.title);
+          console.log('개별 이력 삭제 - 이력 정보:', expToDelete);
+          await deleteExperienceFolder(expToDelete.title);
+        } else {
+          console.log('개별 이력 삭제 - 이력 제목이 없습니다:', expToDelete);
+        }
+  
+        // 로컬 상태에서도 삭제
+        const newExperiences = experiences.filter((_, idx) => idx !== indexToDelete);
+        setExperiences(newExperiences);
+  
+        // 선택된 목록도 갱신 (선택된 이력이 삭제되었을 경우를 대비)
+        setSelected([]);
+  
+        alert('이력이 삭제되었습니다!');
+  
+      } catch (error) {
+        console.error('개별 이력 삭제 오류:', error);
+        const errorMessage = sheetsService.current?.formatErrorMessage(error) || '이력 삭제에 실패했습니다.';
+        alert(errorMessage);
+      } finally {
+        setIsExperienceLoading(false);
+      }
+    }
+  
+    // 파일 다운로드 (Access Token 사용)
+    async function handleDriveFileDownload(file) {
+      if (!driveService.current || !authService.current) return;
+      try {
+        setIsLoading(true);
+        const accessToken = authService.current.getAccessToken();
+        // 구글 문서류(import가 필요한 유형)는 export, 일반 파일은 alt=media
+        const isGoogleDoc = file.mimeType?.includes('application/vnd.google-apps');
+        const exportMap = {
+          'application/vnd.google-apps.document': 'application/vnd.openxmlformats-officedocument.wordprocessingml.document',
+          'application/vnd.google-apps.spreadsheet': 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet',
+          'application/vnd.google-apps.presentation': 'application/vnd.openxmlformats-officedocument.presentationml.presentation',
+          'application/vnd.google-apps.drawing': 'image/png',
+        };
+  
+        const url = isGoogleDoc
+            ? `https://www.googleapis.com/drive/v3/files/${file.id}/export?mimeType=${encodeURIComponent(exportMap[file.mimeType] || 'application/pdf')}`
+            : `https://www.googleapis.com/drive/v3/files/${file.id}?alt=media`;
+  
+        const res = await fetch(url, { headers: { Authorization: `Bearer ${accessToken}` } });
+        if (!res.ok) throw new Error(`다운로드 실패: ${res.status}`);
+        const blob = await res.blob();
+  
+        // 파일명/확장자 보정
+        let filename = file.name;
+        if (isGoogleDoc) {
+          const extMap = {
+            'application/vnd.google-apps.document': '.docx',
+            'application/vnd.google-apps.spreadsheet': '.xlsx',
+            'application/vnd.google-apps.presentation': '.pptx',
+            'application/vnd.google-apps.drawing': '.png',
+          };
+          if (!/\.[a-z0-9]+$/i.test(filename)) filename += (extMap[file.mimeType] || '.pdf');
+        }
+  
+        const objectUrl = URL.createObjectURL(blob);
+        const a = document.createElement('a');
+        a.href = objectUrl;
+        a.download = filename;
+        document.body.appendChild(a);
+        a.click();
+        a.remove();
+        URL.revokeObjectURL(objectUrl);
+      } catch (e) {
+        const msg = driveService.current?.formatErrorMessage?.(e) || e.message || '다운로드 오류';
+        alert(msg);
+      } finally {
+        setIsLoading(false);
+      }
+    }
+  
+  
+    // 파일 업로드 핸들러
+    async function handleDriveFileUpload(event) {
+      const file = event.target.files[0];
+      if (!file || !driveService.current) return;
+  
+      try {
+        setIsUploadLoading(true);
+        await driveService.current.uploadFile(file.name, file, file.type);
+        
+        // 현재 경로를 유지하면서 파일 목록 새로고침
+        const currentParentId = currentPath.length > 0 ? currentPath[currentPath.length - 1].id : null;
+        await loadDriveFiles(currentParentId);
+        
+        alert('파일이 업로드되었습니다!');
+      } catch (error) {
+        const errorMessage = driveService.current?.formatErrorMessage(error) || '파일 업로드에 실패했습니다.';
+        alert(errorMessage);
+      } finally {
+        setIsUploadLoading(false);
+      }
+    }
+  
+    // 파일 삭제 핸들러
+    async function handleDriveFileDelete(fileId, isFromPptHistory = false) {
+      if (!window.confirm('정말로 이 파일을 삭제하시겠습니까?') || !driveService.current) return;
+  
+      try {
+        // 해당 파일 ID를 삭제 중 상태에 추가
+        setDeletingFileIds(prev => new Set([...prev, fileId]));
+        await driveService.current.deleteFile(fileId);
+        
+        if (isFromPptHistory) {
+          // PPT 기록에서 삭제한 경우 PPT 기록 새로고침
+          await loadPptHistory();
+          console.log('PPT 기록이 새로고침되었습니다.');
+        } else {
+          // 드라이브 파일에서 삭제한 경우 드라이브 파일 목록 새로고침
+          const currentParentId = currentPath.length > 0 ? currentPath[currentPath.length - 1].id : null;
+          await loadDriveFiles(currentParentId);
+        }
+        
+        alert('파일이 삭제되었습니다!');
+      } catch (error) {
+        const errorMessage = driveService.current?.formatErrorMessage(error) || '파일 삭제에 실패했습니다.';
+        alert(errorMessage);
+      } finally {
+        // 해당 파일 ID를 삭제 중 상태에서 제거
+        setDeletingFileIds(prev => {
+          const newSet = new Set(prev);
+          newSet.delete(fileId);
+          return newSet;
+        });
+      }
+    }
+  
+    // 구글 시트 데이터 새로고침
+    async function refreshSheetsData() {
+      try {
+        setIsExperienceLoading(true);
+        await loadExperiencesFromSheets();
+      } catch (error) {
+        console.error('시트 데이터 새로고침 오류:', error);
+        alert('데이터 새로고침에 실패했습니다: ' + (error?.message || error));
+      } finally {
+        setIsExperienceLoading(false);
+      }
+    }
+  
+    // 시트 생성
+    async function createSheet() {
+      if (!sheetsService.current || !driveService.current) return;
+  
+      try {
+        setIsSheetLoading(true);
+  
+        // 포트폴리오 이력 폴더 생성 또는 찾기
+        const portfolioFolder = await driveService.current.ensurePortfolioFolder();
+        setPortfolioFolderId(portfolioFolder.id);
+        localStorage.setItem('portfolioFolderId', portfolioFolder.id);
+  
+        // 이미지 폴더도 생성
+        await driveService.current.ensureImageFolder(portfolioFolder.id);
+  
+        // 시트 생성
+        const spreadsheet = await sheetsService.current.createSpreadsheet('포트폴리오 이력', portfolioFolder.id);
+        const newSpreadsheetId = spreadsheet.spreadsheetId;
+  
+        // 상태 업데이트
+        setSpreadsheetId(newSpreadsheetId);
+        localStorage.setItem('spreadsheetId', newSpreadsheetId);
+  
+        // 헤더 설정
+        await sheetsService.current.setupHeaders(newSpreadsheetId);
+  
+        // 파일 목록 새로고침
+        await loadDriveFiles();
+  
+        alert('포트폴리오 시트와 폴더가 생성되었습니다!');
+      } catch (error) {
+        console.error('시트 생성 오류:', error);
+        alert('시트 생성에 실패했습니다: ' + (error?.message || error));
+      } finally {
+        setIsSheetLoading(false);
+      }
+    }
+  
+    // 시트 삭제
+    async function deleteSheet() {
+      if (!spreadsheetId || !driveService.current) return;
+  
+      if (!window.confirm('포트폴리오 시트와 포트폴리오 이력 폴더를 모두 삭제하시겠습니까? 이 작업은 되돌릴 수 없습니다.')) return;
+  
+      try {
+        setIsSheetLoading(true);
+  
+        // 시트 파일 삭제
+        await driveService.current.deleteFile(spreadsheetId);
+  
+        // 포트폴리오 이력 폴더도 삭제
+        if (portfolioFolderId) {
+          try {
+            await driveService.current.deleteFile(portfolioFolderId);
+            console.log('포트폴리오 이력 폴더도 삭제됨');
+          } catch (folderError) {
+            console.warn('포트폴리오 폴더 삭제 실패:', folderError);
+          }
+        }
+  
+        // 상태 초기화
+        setSpreadsheetId(null);
+        localStorage.removeItem('spreadsheetId');
+        setPortfolioFolderId(null);
+        localStorage.removeItem('portfolioFolderId');
+        setExperiences([]);
+  
+        // 강제로 상태 업데이트
+        setTimeout(() => {
+          setSpreadsheetId(null);
+          setPortfolioFolderId(null);
+        }, 100);
+  
+        // 파일 목록 새로고침
+        await loadDriveFiles();
+  
+        alert('포트폴리오 시트와 폴더가 삭제되었습니다!');
+      } catch (error) {
+        console.error('시트 삭제 오류:', error);
+        alert('시트 삭제에 실패했습니다: ' + (error?.message || error));
+      } finally {
+        setIsSheetLoading(false);
+      }
+    }
+  
+    // 뷰 모드 전환
+    async function switchViewMode(mode) {
+      try {
+        setIsViewModeLoading(true);
+        setDriveViewMode(mode);
+        localStorage.setItem('driveViewMode', mode);
+        setCurrentPath([]); // 경로 초기화
+        await loadDriveFiles();
+      } finally {
+        setIsViewModeLoading(false);
+      }
+    }
+  
+    // 드라이브 새로고침
+    async function handleDriveRefresh() {
+      try {
+        setIsRefreshLoading(true);
+        
+        // 현재 경로를 유지하면서 파일 목록 새로고침
+        const currentParentId = currentPath.length > 0 ? currentPath[currentPath.length - 1].id : null;
+        await loadDriveFiles(currentParentId);
+      } finally {
+        setIsRefreshLoading(false);
+      }
+    }
+  
+    // 프레젠테이션 생성
+    async function createPresentation(title, token) {
+      const res = await fetch('https://slides.googleapis.com/v1/presentations', {
+        method: 'POST',
+        headers: {
+          Authorization: `Bearer ${token}`,
+          'Content-Type': 'application/json',
+        },
+        body: JSON.stringify({ title })
+      });
+  
+      const data = await res.json();
+      console.log('Created presentation ID:', data.presentationId);
+      return data.presentationId;
+    }
+  
+    // 슬라이드 추가 (TITLE_AND_BODY)
+    async function addSlide(presentationId, token) {
+      await fetch(`https://slides.googleapis.com/v1/presentations/${presentationId}:batchUpdate`, {
+        method: 'POST',
+        headers: {
+          Authorization: `Bearer ${token}`,
+          'Content-Type': 'application/json',
+        },
+        body: JSON.stringify({
+          requests: [
+            {
+              createSlide: {
+                slideLayoutReference: {
+                  predefinedLayout: 'BLANK'  // 빈 슬라이드로 변경
+                }
+              }
+            }
+          ]
+        })
+      });
+    }
+  
+    // 첫 슬라이드에 제목/부제목 설정 (레이아웃 변경 없이)
+    async function makeTitleAndBody(presId, slideId, token, title, subtitle) {
+      try {
+        // 슬라이드 데이터를 가져와서 텍스트 요소들을 찾기
+        const slideData = await getPresentationData(presId, token);
+        const slide = slideData.slides.find(s => s.objectId === slideId);
+        
+        if (!slide || !slide.pageElements) {
+          console.error('슬라이드 또는 페이지 요소를 찾을 수 없습니다.');
+          return;
+        }
+  
+        console.log('슬라이드 요소들:', slide.pageElements);
+  
+        // 제목과 부제목 요소 찾기
+        let titleElement = null;
+        let subtitleElement = null;
+  
+        for (const element of slide.pageElements) {
+          if (element.shape && element.shape.shapeType === 'TEXT_BOX') {
+            // 첫 번째 텍스트 박스를 제목으로, 두 번째를 부제목으로 사용
+            if (!titleElement) {
+              titleElement = element;
+            } else if (!subtitleElement) {
+              subtitleElement = element;
+            }
+          }
+        }
+  
+        const textRequests = [];
+  
+        // 제목 설정
+        if (titleElement && title) {
+          console.log('제목 설정:', title, '요소 ID:', titleElement.objectId);
+          textRequests.push({
+            insertText: {
+              objectId: titleElement.objectId,
+              insertionIndex: 0,
+              text: title
+            }
+          });
+        }
+  
+        // 부제목 설정
+        if (subtitleElement && subtitle) {
+          console.log('부제목 설정:', subtitle, '요소 ID:', subtitleElement.objectId);
+          textRequests.push({
+            insertText: {
+              objectId: subtitleElement.objectId,
+              insertionIndex: 0,
+              text: subtitle
+            }
+          });
+        }
+  
+        if (textRequests.length > 0) {
+          await fetch(`https://slides.googleapis.com/v1/presentations/${presId}:batchUpdate`, {
+            method: 'POST',
+            headers: {
+              Authorization: `Bearer ${token}`,
+              'Content-Type': 'application/json',
+            },
+            body: JSON.stringify({ requests: textRequests })
+          });
+          console.log('제목과 부제목이 성공적으로 설정되었습니다.');
+        } else {
+          console.log('설정할 텍스트 요소를 찾을 수 없습니다.');
+        }
+  
+      } catch (error) {
+        console.error('makeTitleAndBody 오류:', error);
+      }
+    }
+  
+    async function getPresentationData(presentationId, token) {
+      const res = await fetch(`https://slides.googleapis.com/v1/presentations/${presentationId}`, {
+        method: 'GET',
+        headers: {
+          Authorization: `Bearer ${token}`
+        }
+      });
+  
+      const data = await res.json();
+      console.log('Slides data:', data);
+      return data;
+    }
+  
+    async function updateElementText(presentationId, elementId, newText, token) {
+      const requests = [
+        {
+          insertText: {
+            objectId: elementId,
+            insertionIndex: 0,
+            text: newText
+          }
+        },
+        {
+          deleteText: {
+            objectId: elementId,
+            textRange: { type: 'ALL' }
+          }
+        },
+        {
+          insertText: {
+            objectId: elementId,
+            insertionIndex: 0,
+            text: newText
+          }
+        }
+      ];
+  
+      await fetch(`https://slides.googleapis.com/v1/presentations/${presentationId}:batchUpdate`, {
+        method: 'POST',
+        headers: {
+          Authorization: `Bearer ${token}`,
+          'Content-Type': 'application/json'
+        },
+        body: JSON.stringify({ requests })
+      });
+    }
+  
+    //스타일 변경을 위한 함수 추가
+    async function updateElementStyle(presentationId, elementId, styleObj, token) {
+      const requests = [
+        {
+          updateTextStyle: {
+            objectId: elementId,
+            style: styleObj,
+            fields: Object.keys(styleObj).join(','),
+            textRange: { type: 'ALL' }
+          }
+        }
+      ];
+  
+  
+      await fetch(`https://slides.googleapis.com/v1/presentations/${presentationId}:batchUpdate`, {
+        method: 'POST',
+        headers: {
+          Authorization: `Bearer ${token}`,
+          'Content-Type': 'application/json'
+        },
+        body: JSON.stringify({ requests })
+      });
+    }
+  
+    // 스타일이 적용된 텍스트박스를 슬라이드에 추가
+    async function addStyledTextBoxToSlide(presentationId, slideId, text, token, position = { x: 0, y: 0, width: 300, height: 100 }, style = {}) {
+      try {
+        const objectId = `textbox_${Date.now()}_${Math.random().toString(36).substr(2, 9)}`;
+        
+        const response = await fetch(`https://slides.googleapis.com/v1/presentations/${presentationId}:batchUpdate`, {
+          method: 'POST',
+          headers: {
+            'Authorization': `Bearer ${token}`,
+            'Content-Type': 'application/json'
+          },
+          body: JSON.stringify({
+            requests: [{
+              createShape: {
+                objectId: objectId,
+                shapeType: 'TEXT_BOX',
+                elementProperties: {
+                  pageObjectId: slideId,
+                  size: {
+                    width: { magnitude: position.width, unit: 'PT' },
+                    height: { magnitude: position.height, unit: 'PT' }
+                  },
+                  transform: {
+                    scaleX: 1,
+                    scaleY: 1,
+                    translateX: position.x,
+                    translateY: position.y,
+                    unit: 'PT'
+                  }
+                }
+              }
+            }, {
+              insertText: {
+                objectId: objectId,
+                text: text
+              }
+            }, {
+              updateTextStyle: {
+                objectId: objectId,
+                style: {
+                  bold: style.bold || false,
+                  italic: style.italic || false,
+                  fontSize: style.fontSize || { magnitude: 12, unit: 'PT' },
+                  fontFamily: style.fontFamily || 'Arial',
+                  foregroundColor: style.color || { opaqueColor: { rgbColor: { red: 0, green: 0, blue: 0 } } }
+                },
+                fields: 'bold,italic,fontSize,fontFamily,foregroundColor'
+              }
+            }]
+          })
+        });
+  
+        if (!response.ok) {
+          const errorText = await response.text();
+          throw new Error(`스타일 텍스트박스 추가 실패: ${errorText}`);
+        }
+  
+        const result = await response.json();
+        console.log('스타일 텍스트박스가 슬라이드에 추가되었습니다:', result);
+        return result;
+      } catch (error) {
+        console.error('스타일 텍스트박스 추가 오류:', error);
+        throw error;
+      }
+    }
+  
+    // 텍스트박스를 슬라이드에 추가
+    async function addTextBoxToSlide(presentationId, slideId, text, token, position = { x: 0, y: 0, width: 300, height: 100 }) {
+      try {
+        const objectId = `textbox_${Date.now()}_${Math.random().toString(36).substr(2, 9)}`;
+        
+        const response = await fetch(`https://slides.googleapis.com/v1/presentations/${presentationId}:batchUpdate`, {
+          method: 'POST',
+          headers: {
+            'Authorization': `Bearer ${token}`,
+            'Content-Type': 'application/json'
+          },
+          body: JSON.stringify({
+            requests: [{
+              createShape: {
+                objectId: objectId,
+                shapeType: 'TEXT_BOX',
+                elementProperties: {
+                  pageObjectId: slideId,
+                  size: {
+                    width: { magnitude: position.width, unit: 'PT' },
+                    height: { magnitude: position.height, unit: 'PT' }
+                  },
+                  transform: {
+                    scaleX: 1,
+                    scaleY: 1,
+                    translateX: position.x,
+                    translateY: position.y,
+                    unit: 'PT'
+                  }
+                }
+              }
+            }, {
+              insertText: {
+                objectId: objectId,
+                text: text
+              }
+            }]
+          })
+        });
+  
+        if (!response.ok) {
+          const errorText = await response.text();
+          throw new Error(`텍스트박스 추가 실패: ${errorText}`);
+        }
+  
+        const result = await response.json();
+        console.log('텍스트박스가 슬라이드에 추가되었습니다:', result);
+        return result;
+      } catch (error) {
+        console.error('텍스트박스 추가 오류:', error);
+        throw error;
+      }
+    }
+  
+    // 이미지를 슬라이드에 추가
+    async function addImageToSlide(presentationId, slideId, imageUrl, token, position = { x: 0, y: 0, width: 300, height: 200 }) {
+      try {
+        const response = await fetch(`https://slides.googleapis.com/v1/presentations/${presentationId}:batchUpdate`, {
+          method: 'POST',
+          headers: {
+            Authorization: `Bearer ${token}`,
+            'Content-Type': 'application/json',
+          },
+          body: JSON.stringify({
+            requests: [
+              {
+                createImage: {
+                  objectId: `image_${Date.now()}`,
+                  url: imageUrl,
+                  elementProperties: {
+                    pageObjectId: slideId,
+                    size: {
+                      width: {
+                        magnitude: position.width,
+                        unit: 'PT'
+                      },
+                      height: {
+                        magnitude: position.height,
+                        unit: 'PT'
+                      }
+                    },
+                    transform: {
+                      scaleX: 1,
+                      scaleY: 1,
+                      translateX: position.x,
+                      translateY: position.y,
+                      unit: 'PT'
+                    }
+                  }
+                }
+              }
+            ]
+          })
+        });
+  
+        if (!response.ok) {
+          const errorData = await response.json();
+          console.error('이미지 추가 오류:', errorData);
+          throw new Error(`이미지 추가 실패: ${errorData.error?.message || '알 수 없는 오류'}`);
+        }
+  
+        const result = await response.json();
+        console.log('이미지가 성공적으로 추가되었습니다:', result);
+        return result;
+      } catch (error) {
+        console.error('이미지 추가 중 오류 발생:', error);
+        throw error;
+      }
+    }
+  
+  // API로 텍스트 수정하고 성공하면 로컬 slides 상태도 안전히 갱신하는 헬퍼
+    async function updateElementTextAndLocal(pId, elementId, newText, token) {
+      await updateElementText(pId, elementId, newText, token);
+  
+  
+      setSlides(prevSlides => {
+        return prevSlides.map(slide => {
+          if (!slide.pageElements) return slide;
+          const hasEl = slide.pageElements.some(pe => pe.objectId === elementId);
+          if (!hasEl) return slide;
+  
+  
+          const newPageElements = slide.pageElements.map(pe => {
+            if (pe.objectId !== elementId) return pe;
+            const newShape = {
+              ...pe.shape,
+              text: {
+                ...pe.shape?.text,
+                textElements: [{ textRun: { content: newText } }]
+              }
+            };
+            return { ...pe, shape: newShape };
+          });
+  
+  
+          return { ...slide, pageElements: newPageElements };
+        });
+      });
+    }
+  
+  
+    async function downloadPptxFromDrive(presentationId, token) {
+      const mime = 'application/vnd.openxmlformats-officedocument.presentationml.presentation';
+      const res = await fetch(`https://www.googleapis.com/drive/v3/files/${presentationId}/export?mimeType=${encodeURIComponent(mime)}`, {
+        method: 'GET',
+        headers: {
+          Authorization: `Bearer ${token}`
+        }
+      });
+      
+      if (!res.ok) {
+        const text = await res.text();
+        throw new Error('Export 실패: ' + text);
+      }
+      const blob = await res.blob();
+      const url = URL.createObjectURL(blob);
+      const a = document.createElement('a');
+      a.href = url;
+      a.download = `presentation_${presentationId}.pptx`;
+      document.body.appendChild(a);
+      a.click();
+      a.remove();
+      URL.revokeObjectURL(url);
+    }
+  
+    function findFirstPlaceholder(shapeList, type /* 'TITLE' | 'BODY' */) {
+      for (const el of shapeList) {
+        if (el.shape && el.shape.placeholder && el.shape.placeholder.type === type) {
+          return el.objectId;
+        }
+      }
+      return null;
+    }
+  
+    function getTextFromElement(el) {
+      if (
+          el.shape &&
+          el.shape.text &&
+          el.shape.text.textElements
+      ) {
+        return el.shape.text.textElements
+            .map(te => te.textRun?.content || '')
+            .join('');
+      }
+      return '';
+    }
+  
+    // 템플릿 모달 표시
+    function openTemplateModal(templateName) {
+      setSelectedTemplateForModal(templateName);
+      setShowTemplateModal(true);
+    }
+  
+    // 템플릿 모달에서 사용 버튼 클릭
+    async function handleTemplateUse() {
+      setShowTemplateModal(false);
+      
+      // 인증 서비스를 통해 토큰 확인 및 갱신
+      if (!authService.current) {
+        alert('인증 서비스가 초기화되지 않았습니다. 다시 로그인해주세요.');
+        return;
+      }
+      
+      // 인증 상태 확인
+      const isAuthenticated = await authService.current.isAuthenticated();
+      if (!isAuthenticated) {
+        alert('인증이 필요합니다. 다시 로그인해주세요.');
+        return;
+      }
+      
+      // 액세스 토큰 가져오기
+      const token = await authService.current.getAccessToken();
+      if (!token) {
+        alert('액세스 토큰을 가져올 수 없습니다. 다시 로그인해주세요.');
+        return;
+      }
+      
+      await handleTemplateSelect(selectedTemplateForModal, token);
+    }
+  
+    // 템플릿 모달에서 취소 버튼 클릭
+    function handleTemplateCancel() {
+      setShowTemplateModal(false);
+      setSelectedTemplateForModal(null);
+    }
+  
+    // 템플릿 설명 객체
+    const templateDescriptions = {
+      basic: {
+        name: '기본 템플릿',
+        description: '깔끔하고 전문적인 레이아웃으로 구성된 템플릿입니다. 각 이력사항을 개별 슬라이드로 구성하여 명확하고 간결하게 표현할 수 있습니다.',
+        features: ['깔끔한 디자인', '이력별 개별 슬라이드', '전문적인 레이아웃', '이미지와 텍스트 조화'],
+        previewImages: [
+          `${process.env.PUBLIC_URL}/template/img/sample1.png`,
+          `${process.env.PUBLIC_URL}/template/img/sample2.png`,
+          `${process.env.PUBLIC_URL}/template/img/sample3.png`
+        ]
+      },
+      timeline: {
+        name: '타임라인 템플릿',
+        description: '시간의 흐름에 따라 이력사항을 구성하는 템플릿입니다. 연도별로 정리된 타임라인과 함께 각 이력의 상세 내용을 보여줍니다.',
+        features: ['시간순 구성', '타임라인 시각화', '연도별 정리', '발전 과정 표현'],
+        previewImages: []
+      },
+      grid: {
+        name: '그리드 템플릿',
+        description: '격자 형태로 이력사항을 배치하는 템플릿입니다. 여러 이력을 한눈에 볼 수 있어 비교하기 쉽고 시각적으로 정리된 느낌을 줍니다.',
+        features: ['격자형 레이아웃', '한눈에 보기', '비교 용이', '시각적 정리'],
+        previewImages: []
+      }
+    };
+  
+    // 템플릿 선택 → 프레젠테이션 생성 + 이력 반영
+    async function handleTemplateSelect(templateName, token) {
+      const title = prompt('슬라이드 제목을 입력하세요:', '나의 포트폴리오');
+      if (!title) {
+        alert('제목이 없습니다.');
+        return;
+      }
+  
+      // 현재 날짜를 YYYY-MM-DD 형식으로 생성
+      const currentDate = new Date();
+      const year = currentDate.getFullYear();
+      const month = String(currentDate.getMonth() + 1).padStart(2, '0');
+      const day = String(currentDate.getDate()).padStart(2, '0');
+      const dateString = `${year}-${month}-${day}`;
+      
+      // PPT 표지용 제목 (날짜와 중복표시 제거)
+      const cleanTitle = title.replace(/\s+\d{4}-\d{2}-\d{2}$/, '').replace(/_\d+$/, '');
+      
+      // 기본 파일명 생성 (날짜 포함)
+      const baseFileName = `${title} ${dateString}`;
+      
+      // PPT 폴더에서 기존 PPT 파일들 조회하여 중복 확인
+      let finalFileName = baseFileName;
+      try {
+        const portfolioFolder = await driveService.current.ensurePortfolioFolder();
+        const pptFolder = await driveService.current.findFolder('PPT', portfolioFolder.id);
+        
+        if (pptFolder) {
+          const existingFiles = await driveService.current.getFilesInFolder(pptFolder.id);
+          
+          // PPT 파일만 필터링 (Google Slides 파일)
+          const pptFiles = existingFiles.filter(file => 
+            file.mimeType === 'application/vnd.google-apps.presentation'
+          );
+          
+          // 순차적 번호가 포함된 파일명 생성
+          finalFileName = driveService.current.generateSequentialFileName(baseFileName, pptFiles);
+          console.log(`원본 PPT 파일명: ${baseFileName} → 최종 파일명: ${finalFileName}`);
+        } else {
+          console.log('PPT 폴더가 없어서 기본 파일명 사용');
+        }
+      } catch (error) {
+        console.warn('PPT 파일명 중복 확인 실패, 기본 파일명 사용:', error);
+      }
+  
+      // 토큰 확보 보장
+      let currentToken = accessToken;
+      if (!currentToken) {
+        try {
+          // 인증 상태 확인
+          const isAuthenticated = await authService.current.isAuthenticated();
+          if (!isAuthenticated) {
+            alert('인증이 필요합니다. 다시 로그인해주세요.');
+            return;
+          }
+          
+          currentToken = await authService.current.getAccessToken();
+          if (!currentToken) {
+            alert('액세스 토큰을 가져올 수 없습니다. 다시 로그인해주세요.');
+            return;
+          }
+          setAccessToken(currentToken);
+        } catch (error) {
+          console.error('토큰 가져오기 실패:', error);
+          alert('인증이 필요합니다. 다시 로그인해주세요.');
+          return;
+        }
+      }
+  
+      setSelectedTemplate(templateName);
+      setIsPptCreating(true); // PPT 생성 시작
+  
+      try {
+        // 1) 프레젠테이션 생성 (중복 확인된 최종 파일명으로)
+        const presId = await createPresentation(finalFileName, currentToken);
+        setPresentationId(presId);
+  
+        // 1-1) PPT 파일을 PPT 폴더로 이동
+        try {
+          const portfolioFolder = await driveService.current.ensurePortfolioFolder();
+          const pptFolder = await driveService.current.findFolder('PPT', portfolioFolder.id);
+          
+          if (pptFolder) {
+            const gapiClient = authService.current.getAuthenticatedGapiClient();
+            
+            // PPT 파일을 PPT 폴더로 이동
+            await gapiClient.drive.files.update({
+              fileId: presId,
+              addParents: pptFolder.id,
+              removeParents: 'root'
+            });
+            console.log('PPT 파일이 PPT 폴더로 이동되었습니다.');
+          } else {
+            console.warn('PPT 폴더를 찾을 수 없습니다. 포트폴리오 폴더에 저장됩니다.');
+          }
+        } catch (moveError) {
+          console.warn('PPT 파일 폴더 이동 실패:', moveError);
+          // 폴더 이동 실패해도 PPT 생성은 성공으로 처리
+        }
+  
+        // 2) 첫 슬라이드 레이아웃 보정 및 제목/부제목 설정
+        let data = await getPresentationData(presId, currentToken);
+        if (data.slides?.length > 0) {
+          // 구글 계정에서 사용자 이름 가져오기
+          let userName = '사용자';
+          try {
+            userName = await authService.current.getUserInfo();
+          } catch (error) {
+            console.warn('사용자 이름 가져오기 실패, 기본값 사용:', error);
+          }
+          
+          await makeTitleAndBody(presId, data.slides[0].objectId, currentToken, cleanTitle, userName);
+        }
+  
+        // 3) 템플릿별 슬라이드 추가
+        if (templateName === 'basic') {
+          for (let i = 0; i < selectedExperiences.length; i++) {
+            await addSlide(presId, currentToken);
+          }
+        } else if (templateName === 'timeline') {
+          await addSlide(presId, currentToken);
+          await addSlide(presId, currentToken);
+          for (let i = 0; i < selectedExperiences.length; i++) {
+            await addSlide(presId, currentToken);
+          }
+        } else if (templateName === 'grid') {
+          await addSlide(presId, currentToken);
+          await addSlide(presId, currentToken);
+        }
+  
+        // 4) 최신 데이터 가져오기
+        data = await getPresentationData(presId, currentToken);
+        const slidesArr = data.slides || [];
+  
+        // 5) 텍스트 채우기
+        if (slidesArr[0]) {
+          const s0 = slidesArr[0];
+          const titleShapeId = findFirstPlaceholder(s0.pageElements, 'TITLE');
+          const bodyShapeId  = findFirstPlaceholder(s0.pageElements, 'BODY');
+  
+          // 표지 제목: 원본 제목 (날짜 제외)
+          if (titleShapeId) await updateElementText(presId, titleShapeId, title, currentToken);
+          
+          // 표지 본문: 구글 계정 이름
+          if (bodyShapeId) {
+            const userName = await getGoogleAccountName();
+            await updateElementText(presId, bodyShapeId, userName, currentToken);
+          }
+        }
+  
+        let idx = 1;
+        for (const exp of selectedExperiences) {
+          if (!slidesArr[idx]) break;
+          const s = slidesArr[idx];
+  
+          // 제목, 기간, 설명을 위한 새로운 텍스트박스들을 생성
+          try {
+            // 1. 제목 텍스트박스 (제목 스타일 적용)
+            await addStyledTextBoxToSlide(presId, s.objectId, exp.title, token, {
+              x: 50,   // 왼쪽에서 50pt
+              y: 50,   // 상단에서 50pt
+              width: 400,
+              height: 60
+            }, {
+              bold: true,
+              fontSize: { magnitude: 24, unit: 'PT' },
+              fontFamily: 'Arial',
+              color: { opaqueColor: { rgbColor: { red: 0, green: 0, blue: 0 } } }
+            });
+            console.log(`제목 텍스트박스가 슬라이드 ${idx}에 추가되었습니다:`, exp.title);
+  
+            // 2. 기간 텍스트박스 (일반 스타일, 가로 길이 증가)
+            await addTextBoxToSlide(presId, s.objectId, exp.period, token, {
+              x: 50,   // 왼쪽에서 50pt
+              y: 100,  // 제목 아래 간격을 줄임 (120 → 100)
+              width: 350,  // 가로 길이를 늘림 (200 → 350)
+              height: 40
+            });
+            console.log(`기간 텍스트박스가 슬라이드 ${idx}에 추가되었습니다:`, exp.period);
+  
+            // 3. 설명 텍스트박스 (일반 스타일, 위치 조정)
+            await addTextBoxToSlide(presId, s.objectId, exp.description, token, {
+              x: 50,   // 왼쪽에서 50pt
+              y: 150,  // 기간 아래 간격을 줄임 (170 → 150)
+              width: 300,
+              height: 80
+            });
+            console.log(`설명 텍스트박스가 슬라이드 ${idx}에 추가되었습니다:`, exp.description);
+  
+          } catch (textBoxError) {
+            console.error(`슬라이드 ${idx}에 텍스트박스 추가 실패:`, textBoxError);
+            // 텍스트박스 추가 실패해도 PPT 생성은 계속 진행
+          }
+          
+          // 이미지가 있는 경우 슬라이드에 추가
+          if (exp.imageUrls && exp.imageUrls.length > 0) {
+            try {
+              // 모든 이미지를 추가 (세로로 일렬 배치)
+              const imageCount = exp.imageUrls.length;
+              const imageWidth = 250;
+              const imageHeight = 150;
+              const imageSpacing = 20;
+              const startX = 400; // 텍스트 영역 오른쪽
+              const startY = 100; // 상단에서 100pt
+              
+              for (let i = 0; i < imageCount; i++) {
+                const imageUrl = exp.imageUrls[i];
+                
+                // 이미지 위치 계산 (세로로 일렬 배치)
+                const imagePosition = {
+                  x: startX,
+                  y: startY + i * (imageHeight + imageSpacing),
+                  width: imageWidth,
+                  height: imageHeight
+                };
+                
+                await addImageToSlide(presId, s.objectId, imageUrl, token, imagePosition);
+                console.log(`이미지 ${i + 1}/${imageCount}가 슬라이드 ${idx}에 추가되었습니다:`, imageUrl);
+              }
+            } catch (imageError) {
+              console.error(`슬라이드 ${idx}에 이미지 추가 실패:`, imageError);
+              // 이미지 추가 실패해도 PPT 생성은 계속 진행
+            }
+          }
+          
+          idx++;
+        }
+  
+        // 6) 최종 상태 반영
+        const refreshed = await getPresentationData(presId, currentToken);
+        setSlides(refreshed.slides || []);
+        
+        // 7) PPT 기록 새로고침
+        await loadPptHistory();
+        
+        alert('PPT가 생성되었습니다.');
+        setActiveSection('editor');
+      } catch (error) {
+        console.error('PPT 생성 오류:', error);
+        alert('PPT 생성에 실패했습니다: ' + (error?.message || error));
+      } finally {
+        setIsPptCreating(false); // PPT 생성 완료 (성공/실패 관계없이)
+      }
+    }
+  
+    // 폴더 진입
+    async function enterFolder(folderId, folderName) {
+      try {
+        setIsViewModeLoading(true);
+        
+        // 포트폴리오 폴더인지 확인
+        if (folderName === '포트폴리오 이력') {
+          // 포트폴리오 폴더 ID가 설정되어 있으면 해당 ID와 비교
+          if (portfolioFolderId && folderId === portfolioFolderId) {
+            // 포트폴리오 폴더인 경우 작업영역을 포트폴리오 폴더로 전환
+            setDriveViewMode('portfolio');
+            setCurrentPath([]); // 경로 초기화
+            await loadDriveFiles(); // 포트폴리오 폴더 내용 로드
+          } else {
+            // 포트폴리오 폴더 ID가 설정되지 않은 경우, 폴더 이름으로만 확인
+            // 이 경우 포트폴리오 폴더 ID를 설정하고 작업영역 전환
+            setPortfolioFolderId(folderId);
+            localStorage.setItem('portfolioFolderId', folderId);
+            setDriveViewMode('portfolio');
+            setCurrentPath([]); // 경로 초기화
+            await loadDriveFiles(); // 포트폴리오 폴더 내용 로드
+          }
+        } else {
+          // 일반 폴더인 경우 기존 로직 유지
+          setCurrentPath(prev => [...prev, { id: folderId, name: folderName }]);
+          await loadDriveFiles(folderId);
+        }
+      } finally {
+        setIsViewModeLoading(false);
+      }
+    }
+  
+    // 뒤로가기
+    async function goBack() {
+      if (currentPath.length === 0) return;
+  
+      try {
+        setIsViewModeLoading(true);
+        const newPath = currentPath.slice(0, -1);
+        setCurrentPath(newPath);
+  
+        if (newPath.length === 0) {
+          // 루트로 돌아가기
+          await loadDriveFiles();
+        } else {
+          // 이전 폴더로 돌아가기
+          const parentFolderId = newPath[newPath.length - 1].id;
+          await loadDriveFiles(parentFolderId);
+        }
+      } finally {
+        setIsViewModeLoading(false);
+      }
+    }
+  
+    // 파일 다운로드
+    async function downloadFile(file) {
+      if (!driveService.current) return;
+  
+      try {
+        // 구글 문서 파일인지 확인 (Google Docs, Sheets, Slides 등)
+        const isGoogleDoc = file.mimeType && file.mimeType.includes('application/vnd.google-apps');
+  
+        if (isGoogleDoc) {
+          // 구글 문서 파일은 export API 사용
+          await downloadGoogleDoc(file);
+        } else {
+          // 일반 파일은 직접 다운로드
+          await downloadRegularFile(file);
+        }
+      } catch (error) {
+        console.error('파일 다운로드 오류:', error);
+        alert('파일 다운로드에 실패했습니다: ' + (error?.message || error));
+      }
+    }
+  
+    // 구글 문서 파일 다운로드
+    async function downloadGoogleDoc(file) {
+      try {
+        const accessToken = authService.current?.getAccessToken();
+        if (!accessToken) {
+          throw new Error('인증 토큰이 없습니다.');
+        }
+  
+        // MIME 타입에 따른 export 형식 결정
+        let exportMimeType;
+        let fileExtension;
+  
+        switch (file.mimeType) {
+          case 'application/vnd.google-apps.spreadsheet':
+            exportMimeType = 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet';
+            fileExtension = '.xlsx';
+            break;
+          case 'application/vnd.google-apps.document':
+            exportMimeType = 'application/vnd.openxmlformats-officedocument.wordprocessingml.document';
+            fileExtension = '.docx';
+            break;
+          case 'application/vnd.google-apps.presentation':
+            exportMimeType = 'application/vnd.openxmlformats-officedocument.presentationml.presentation';
+            fileExtension = '.pptx';
+            break;
+          case 'application/vnd.google-apps.drawing':
+            exportMimeType = 'image/png';
+            fileExtension = '.png';
+            break;
+          default:
+            exportMimeType = 'application/pdf';
+            fileExtension = '.pdf';
+        }
+  
+        const exportUrl = `https://www.googleapis.com/drive/v3/files/${file.id}/export?mimeType=${exportMimeType}`;
+  
+        const response = await fetch(exportUrl, {
+          headers: {
+            'Authorization': `Bearer ${accessToken}`,
+          },
+        });
+  
+        if (!response.ok) {
+          throw new Error(`Export 실패: ${response.status} ${response.statusText}`);
+        }
+  
+        const blob = await response.blob();
+        const downloadUrl = window.URL.createObjectURL(blob);
+  
+        const link = document.createElement('a');
+        link.href = downloadUrl;
+        link.download = file.name + fileExtension;
+        document.body.appendChild(link);
+        link.click();
+        document.body.removeChild(link);
+  
+        // 메모리 정리
+        window.URL.revokeObjectURL(downloadUrl);
+  
+      } catch (error) {
+        console.error('구글 문서 다운로드 오류:', error);
+        throw error;
+      }
+    }
+  
+    // 일반 파일 다운로드
+    async function downloadRegularFile(file) {
+      try {
+        const accessToken = authService.current?.getAccessToken();
+        if (!accessToken) {
+          throw new Error('인증 토큰이 없습니다.');
+        }
+  
+        const downloadUrl = `https://www.googleapis.com/drive/v3/files/${file.id}?alt=media`;
+  
+        const response = await fetch(downloadUrl, {
+          headers: {
+            'Authorization': `Bearer ${accessToken}`,
+          },
+        });
+  
+        if (!response.ok) {
+          throw new Error(`다운로드 실패: ${response.status} ${response.statusText}`);
+        }
+  
+        const blob = await response.blob();
+        const objectUrl = window.URL.createObjectURL(blob);
+  
+        const link = document.createElement('a');
+        link.href = objectUrl;
+        link.download = file.name;
+        document.body.appendChild(link);
+        link.click();
+        document.body.removeChild(link);
+  
+        // 메모리 정리
+        window.URL.revokeObjectURL(objectUrl);
+  
+      } catch (error) {
+        console.error('일반 파일 다운로드 오류:', error);
+        throw error;
+      }
+    }
+  
+    // 뷰 모드 변경 시 파일 목록 다시 로드
+    useEffect(() => {
+      if (isDriveInitialized) {
+        loadDriveFiles();
+      }
+    }, [driveViewMode, portfolioFolderId]);
+  
+    // 포트폴리오 폴더 ID 설정
+    async function setPortfolioFolder() {
+      if (!driveService.current) return;
+  
+      try {
+        // 기존 폴더가 있는지 확인만 (생성하지 않음)
+        const portfolioFolder = await driveService.current.findFolder('포트폴리오 이력');
+        if (portfolioFolder) {
+          setPortfolioFolderId(portfolioFolder.id);
+          localStorage.setItem('portfolioFolderId', portfolioFolder.id);
+          console.log('기존 포트폴리오 폴더 ID 설정됨:', portfolioFolder.id);
+        } else {
+          console.log('포트폴리오 폴더가 없습니다. 시트 생성 시 함께 생성됩니다.');
+        }
+      } catch (error) {
+        console.error('포트폴리오 폴더 확인 오류:', error);
+      }
+    }
+  
+    // 드라이브 섹션이 활성화될 때 포트폴리오 폴더 ID 설정
+    useEffect(() => {
+      if (activeSection === 'drive' && isDriveInitialized && !portfolioFolderId) {
+        setPortfolioFolder();
+      }
+    }, [activeSection, isDriveInitialized, portfolioFolderId]);
+  
+    // 포트폴리오 폴더 ID가 변경될 때 localStorage 업데이트
+    useEffect(() => {
+      if (portfolioFolderId) {
+        localStorage.setItem('portfolioFolderId', portfolioFolderId);
+      }
+    }, [portfolioFolderId]);
+  
+    // 마이페이지 섹션이 활성화될 때 PPT 기록과 이력 목록 로드
+    useEffect(() => {
+      if (activeSection === 'myPage' && isDriveInitialized) {
+        loadPptHistory();
+        // 이력 목록도 자동으로 새로고침
+        if (isSheetsInitialized) {
+          loadExperiencesFromSheets();
+        }
+      }
+    }, [activeSection, isDriveInitialized, isSheetsInitialized]);
+  
+    // 페이지 로드 시 로그인 상태 복원 및 초기화
+    useEffect(() => {
+      const initializeApp = async () => {
+        try {
+          console.log('앱 초기화 시작...');
+  
+          // localStorage에서 로그인 상태 확인
+          const savedLoginState = localStorage.getItem('isLoggedIn');
+          const savedSpreadsheetId = localStorage.getItem('spreadsheetId');
+  
+          if (savedLoginState === 'true' && savedSpreadsheetId) {
+            console.log('저장된 로그인 상태 발견, 서비스 초기화 시작...');
+  
+            // 로그인 상태를 먼저 설정
+            setIsLoggedIn(true);
+            setSpreadsheetId(savedSpreadsheetId);
+  
+            // 통합 인증 시스템 초기화
+            await initializeGoogleAuth();
+          } else {
+            console.log('저장된 로그인 상태가 없습니다.');
+            // 로그인 상태가 없으면 명시적으로 false로 설정
+            setIsLoggedIn(false);
+          }
+        } catch (error) {
+          console.error('앱 초기화 오류:', error);
+          // 초기화 실패 시 로그인 상태를 false로 설정
+          setIsLoggedIn(false);
+          setAuthStatus('error');
+        }
+      };
+  
+      initializeApp();
+    }, []);
+  
+  
+    // GIS 기반 로그인 버튼 렌더링
+    useEffect(() => {
+      if (!isLoggedIn) {
+        // GIS 기반 로그인 버튼 렌더링
+        const googleSignInDiv = document.getElementById('googleSignInDiv');
+        if (googleSignInDiv) {
+          googleSignInDiv.innerHTML = `
+            <button 
+              id="gisLoginBtn"
+              style="
+                background: #4285f4; 
+                color: white; 
+                border: none; 
+                padding: 12px 24px; 
+                border-radius: 4px; 
+                font-size: 16px; 
+                cursor: pointer;
+                width: 300px;
+                display: flex;
+                align-items: center;
+                justify-content: center;
+                gap: 8px;
+              "
+            >
+              <svg width="18" height="18" viewBox="0 0 24 24">
+                <path fill="currentColor" d="M22.56 12.25c0-.78-.07-1.53-.2-2.25H12v4.26h5.92c-.26 1.37-1.04 2.53-2.21 3.31v2.77h3.57c2.08-1.92 3.28-4.74 3.28-8.09z"/>
+                <path fill="currentColor" d="M12 23c2.97 0 5.46-.98 7.28-2.66l-3.57-2.77c-.98.66-2.23 1.06-3.71 1.06-2.86 0-5.29-1.93-6.16-4.53H2.18v2.84C3.99 20.53 7.7 23 12 23z"/>
+                <path fill="currentColor" d="M5.84 14.09c-.22-.66-.35-1.36-.35-2.09s.13-1.43.35-2.09V7.07H2.18C1.43 8.55 1 10.22 1 12s.43 3.45 1.18 4.93l2.85-2.22.81-.62z"/>
+                <path fill="currentColor" d="M12 5.38c1.62 0 3.06.56 4.21 1.64l3.15-3.15C17.45 2.09 14.97 1 12 1 7.7 1 3.99 3.47 2.18 7.07l3.66 2.84c.87-2.6 3.3-4.53 6.16-4.53z"/>
+              </svg>
+              구글로 로그인
+            </button>
+          `;
+  
+          // 버튼에 이벤트 리스너 추가
+          const loginBtn = document.getElementById('gisLoginBtn');
+          if (loginBtn) {
+            loginBtn.addEventListener('click', handleGISLogin);
+          }
+        }
+      }
+    }, [isLoggedIn]);
+  
+    // 로딩 상태에 따른 버튼 업데이트
+    useEffect(() => {
+      const loginBtn = document.getElementById('gisLoginBtn');
+      if (loginBtn) {
+        if (isLoading) {
+          loginBtn.innerHTML = `
+            <div class="spinner-border spinner-border-sm" role="status" style="width: 18px; height: 18px;">
+              <span class="visually-hidden">로딩중...</span>
+            </div>
+            로그인 중...
+          `;
+          loginBtn.disabled = true;
+          loginBtn.style.cursor = 'not-allowed';
+          loginBtn.style.opacity = '0.7';
+        } else {
+          loginBtn.innerHTML = `
+            <svg width="18" height="18" viewBox="0 0 24 24">
+              <path fill="currentColor" d="M22.56 12.25c0-.78-.07-1.53-.2-2.25H12v4.26h5.92c-.26 1.37-1.04 2.53-2.21 3.31v2.77h3.57c2.08-1.92 3.28-4.74 3.28-8.09z"/>
+              <path fill="currentColor" d="M12 23c2.97 0 5.46-.98 7.28-2.66l-3.57-2.77c-.98.66-2.23 1.06-3.71 1.06-2.86 0-5.29-1.93-6.16-4.53H2.18v2.84C3.99 20.53 7.7 23 12 23z"/>
+              <path fill="currentColor" d="M5.84 14.09c-.22-.66-.35-1.36-.35-2.09s.13-1.43.35-2.09V7.07H2.18C1.43 8.55 1 10.22 1 12s.43 3.45 1.18 4.93l2.85-2.22.81-.62z"/>
+              <path fill="currentColor" d="M12 5.38c1.62 0 3.06.56 4.21 1.64l3.15-3.15C17.45 2.09 14.97 1 12 1 7.7 1 3.99 3.47 2.18 7.07l3.66 2.84c.87-2.6 3.3-4.53 6.16-4.53z"/>
+            </svg>
+            구글로 로그인
+          `;
+          loginBtn.disabled = false;
+          loginBtn.style.cursor = 'pointer';
+          loginBtn.style.opacity = '1';
+        }
+      }
+    }, [isLoading]);
+  
+    // ESC 키로 이미지 모달 닫기
+    useEffect(() => {
+      const handleEscKey = (event) => {
+        if (event.key === 'Escape' && showImageModal) {
+          closeImageModal();
+        }
+      };
+  
+      if (showImageModal) {
+        document.addEventListener('keydown', handleEscKey);
+      }
+  
+      return () => {
+        document.removeEventListener('keydown', handleEscKey);
+      };
+    }, [showImageModal]);
+
+    // 모든 상태와 함수들을 반환
+    return {
+      // 상태들
+      isLoggedIn,
+      activeSection,
+      authStatus,
+      experiences,
+      selected,
+      spreadsheetId,
+      isSheetsInitialized,
+      isDriveInitialized,
+      driveFiles,
+      isLoading,
+      isSheetLoading,
+      isExperienceLoading,
+      isDriveLoading,
+      isUploadLoading,
+      isRefreshLoading,
+      isDeleteLoading,
+      editingIndex,
+      deletingFileIds,
+      isViewModeLoading,
+      isPptCreating,
+      currentPath,
+      selectedImages,
+      imagePreviews,
+      showImageModal,
+      pptHistory,
+      selectedImageForModal,
+      imageLoadingStates,
+      selectedExperience,
+      showTemplateModal,
+      selectedTemplateForModal,
+      showExperienceModal,
+      showModal,
+      form,
+      formRef,
+      driveViewMode,
+      portfolioFolderId,
+      presentationId,
+      slides,
+      accessToken,
+      selectedExperiences,
+      templateDescriptions,
+      // 함수들
+      showSection,
+      logout,
+      showAddExperienceModal,
+      selectAllExperiences,
+      deleteSelectedExperiences,
+      refreshSheetsData,
+      openExperienceModal,
+      openImageModal,
+      setImageLoadingState,
+      retryImageLoad,
+      toggleSelect,
+      setSelectedExperiences,
+      openTemplateModal,
+      handleTemplateCancel,
+      handleTemplateUse,
+      closeModal,
+      saveExperience,
+      closeImageModal,
+      closeExperienceModal,
+      showEditExperienceModal,
+      deleteIndividualExperience,
+      handleImageSelect,
+      handleDroppedFiles,
+      removeImage,
+      formatPeriod,
+      createSheet,
+      deleteSheet,
+      switchViewMode,
+      goBack,
+      handleDriveFileUpload,
+      handleDriveRefresh,
+      handleDriveFileDelete,
+      enterFolder,
+      downloadFile,
+      loadPptHistory,
+      loadPptForEdit,
+      getTextFromElement,
+      updateElementTextAndLocal,
+      updateElementStyle,
+      // 상태 setter 함수들
+      setActiveSection,
+      setSlides,
+      setForm,
+      setSelectedImageForModal,
+      setShowImageModal
+    };
+}
+
+export { useAppLogic };
